Add a column of institution names

Add a column of character values with institution names (or labels) using student ID as the join-by variable. Obtains the information from the MIDFIELD term data table or equivalent. In the MIDFIELD practice data, the labels are de-identified.

add_institution(dframe, midfield_term)

Arguments

dframe
Data frame with required variable mcid.

midfield_term
MIDFIELD term data table or equivalent with required variables mcid, institutiion, term.

Value

A data.table with the following properties:

Rows are not modified.

Column institution is added.

Grouping structures are not preserved.

Details

If a student is associated with more than one institution, the institution at which they completed the most terms is returned. An existing column with the same name as the added column is overwritten.

See also

Examples

# Extract a column of IDs from student
id <- toy_student[, .(mcid)]

# Add institutions from term
DT1 <- add_institution(id, midfield_term = toy_term)
head(DT1)

#>           mcid   institution
#> 1: MID25783939 Institution M
#> 2: MID25784402 Institution M
#> 3: MID25805538 Institution M
#> 4: MID25808099 Institution M
#> 5: MID25816437 Institution M
#> 6: MID25826223 Institution M

nrow(DT1)

#> [1] 100

# Will overwrite institution column if present
DT2 <- add_institution(DT1, midfield_term = toy_term)
head(DT2)

#>           mcid   institution
#> 1: MID25783939 Institution M
#> 2: MID25784402 Institution M
#> 3: MID25805538 Institution M
#> 4: MID25808099 Institution M
#> 5: MID25816437 Institution M
#> 6: MID25826223 Institution M

nrow(DT2)

#> [1] 100