Stickiness case data

An data frame of stickiness data grouped by program and student race/ethnicity and sex.

case_stickiness

Format

A data frame (tibble) with 32 observations and 6 variables.

program

Instructional programs selected for study.

race

Student race and sex from midfieldstudents data.

sex

Student sex from midfieldstudents data.

ever

Number of students ever enrolled in a program. Grouped and summarized from midfieldterms data.

grad

Number of students graduating from the program. Grouped and summarized from midfielddegrees data.

stick

Program stickiness: the ratio of the number of students graduating from a program to the number ever enrolled in the program.

Source

midfielddata package

Details

For Civil, Electrical, Industrial, and Mechanical Engineering, the number of students ever enrolled, the number of graduates, and the resulting stickiness.