Select 6-digit CIPs and add program labels

From a CIP data frame, select the columns with the 6-digit codes and names and add a new character variable named program.

cip6_select(
  data,
  program = NULL,
  ...,
  cip6 = "cip6",
  cip6name = "cip6name",
  cip4name = "cip4name",
  cip2name = "cip2name"
)

Arguments

data
Data frame of CIP codes and names with the same structure as midfieldr cip.

program
Character variable. There are four options:

If program = NULL, the default, the 4-digit CIP program names are assigned to the new column.

If program = "named_series", then the series name is assigned to the new column.

If one of three strings ("cip2name", "cip4name", or "cip6name"), then the 2-digit, 4-digit, or 6-digit CIP program names are assigned to the new column.

Any other string fills the new column.

...
Not used for values, forces later arguments to bind by name.

cip6
Optional argument, the column name in quotes of the 6-digit CIP code variable in data. Default is "cip6".

cip6name
Optional argument, the column name in quotes of the 6-digit CIP program name variable in data. Default is "cip6name".

cip4name
Optional argument, the column name in quotes of the 4-digit CIP program name variable in data. Default is "cip4name".

cip2name
Optional argument, the column name in quotes of the 2-digit CIP program name variable in data. Default is "cip2name".

Value

A data frame with the 6-digit code and name columns from the input data frame with an added program variable.

Details

The 2-digit and 4-digit codes and names are silently dropped.

Assigning a custom label to a program or a group of programs provides the option of grouping, summarizing, and joining data in ways not afforded by the default CIP data. We can also create program names that are less verbose than the default names.

See also

Examples

# Extract the Philosophy and Religion programs and label
x <- cip_filter(cip, keep_any = "^38")
y <- cip6_select(x, program = "Phil-Relig")

# With no program argument, the 6-digit CIP names are assigned
y <- cip6_select(x)

# Argument options include
y <- cip6_select(x, program = "cip6name")
y <- cip6_select(x, program = "cip4name")
y <- cip6_select(x, program = "cip2name")