Determine the first post-FYE program by student ID

Subset a vector of student IDs to retain students ever enrolled in FYE and return the CIP code of their first post-FYE program.

cip_after_fye(id = NULL, ..., data_students = NULL, data_terms = NULL)

Arguments

id
character vector of student IDs

...
not used for values. Forces optional arguments to be usable by name only.

data_students
data frame of student attributes at matriculation

data_terms
data frame of term attributes

Value

data.frame with the following properties:

Rows for which student ID were ever enrolled in an FYE program

Columns id and cip6

Grouping structures, if any, are not preserved

Data frame extensions tbl or data.table are preserved

Details

Computing some persistence metrics, e.g., graduation rate, is complicated for students required to enroll in first-year engineering (FYE) programs. In such cases, we must predict the students' "starting majors." cip_after_fye() takes the first step in this process by identifying the programs in which students enroll immediately following their final FYE term.

FYE

First year engineering program, typically a common first year curriculum that is a prerequisite for declaring an engineering major

starting major

The degree-granting engineering program---for example, Civil Engineering, Electrical Engineering, Mechanical Engineering, etc.---that we predict the student would have declared had they not been required to enroll in FYE.

The function accesses data_students (midfieldstudents or its equivalent) to restrict the IDs to those enrolled in an FYE program at matriculation. It also accesses data_terms (midfieldterms or its equivalent) to determine the 6-digit CIP code of a student's first post-FYE program. For students with no program after FYE, the CIP is NA.

See also

Examples

# placeholder