Filter CIP data

Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.

cip_filter(series = NULL, data = NULL)

Arguments

series
The conditions used to filter the data. A character vector (or coercible to one) of any combination of 2, 4, or 6-digit CIP codes. If NULL, data is returned unchanged.

data
A data frame of character variables (or coercible to one). Required variables are CIP2, CIP4, and CIP6. If NULL, data is the midfieldr cip dataset.

Value

A data frame: data filtered by the conditions in series.

Details

The purpose of cip_filter() is to obtain the CIP codes of specific programs and use the results to obtain the records of students in those programs.

Examples

cip_filter(series = c("490101", "490205"))
cip_filter(series = c("^4901", "^4902"))
cip_filter(series = c("^49", "^99"))
cip_filter(series = c("^54", "^4901", "490205"))
cip_filter(series = seq(540102, 540108))