Condition FYE data for multiple imputation

Filter first-year-engineering (FYE) students and prepare variables for predicting unknown starting majors. The prepared variables are institution, race, and sex (predictors) and cip6 (with missing values to be imputed). Obtains the requisite information from the MIDFIELD term data table or equivalent. Returns a data frame formatted for multiple imputation using the mice package.

condition_fye(dframe, midfield_term, ..., fye_codes = NULL)

Arguments

dframe
Data frame of all degree-seeking engineering students in the database, with required variables mcid, race, and sex.

midfield_term
MIDFIELD term data table or equivalent with required variables mcid, institution, term, and cip6.

...
Not used, forces later arguments to be used by name.

fye_codes
Optional character vector of 6-digit CIP codes to identify FYE programs, default 140102. Codes must be 6-digit strings of numbers; regular expressions are prohibited. Non-engineering codes---those that do not start with 14"---are ignored.

Value

A data.table with the following properties:

One row for every FYE student.

Columns for ID, institution, race, sex, and CIP code, all except ID converted to factors. Additional columns in dframe are dropped.

Grouping structures are not preserved.

Details

Some US institutions have first year engineering (FYE) programs, typically a common first year curriculum that is a prerequisite for declaring an engineering major. FYE programs are problematic for some persistence metrics. For example, in the conventional graduation rate metric, the only students who count as graduates are those who complete the same program in which they were admitted. But students cannot graduate in FYE; instead, upon completing FYE, they transition to a degree-granting engineering program.

Therefore, to include FYE students in any persistence metric requiring a degree-granting "starting" program, we have to predict the engineering program the FYE student would have declared had they not been required to enroll in FYE. predict_fye() sorts students into two categories:

Students who complete an FYE and declare an engineering major. This is the easy case--at the student's first opportunity, they enrolled in an engineering program of their choosing. We use that program as the predicted starting program.

Students who, after FYE, do not declare an engineering major. This is the more complicated case---the data provide no information regarding what engineering program the student would have declared originally had the institution not required them to enroll in FYE. For these students, we treat the starting program as missing data be predicted using multiple imputation.

The function extracts all terms for all FYE students from midfield_term. In cases where students enter FYE, change programs, and re-enter FYE, only the first group of FYE terms is considered. Any programs before FYE are ignored. The first (if any) post-FYE program is identified. If the program is in engineering, the CIP is retained as their predicted starting major. If not, the CIP is replaced with NA to be treated as missing data for the imputation.

Lastly, the predictor variables (institution, race, sex) and the imputed variable (cip6) are converted to unordered factors. The resulting data frame is ready for use as input for the mice package.

See also

Examples

# Using toy data
DT <- toy_student[, .(mcid, race, sex)]
condition_fye(dframe = DT, midfield_term = toy_term)

#>            mcid   institution            race    sex   cip6
#>  1: MID26060301 Institution C           Asian Female   <NA>
#>  2: MID25995980 Institution C Hispanic/Latinx Female   <NA>
#>  3: MID25997636 Institution C Hispanic/Latinx Female   <NA>
#>  4: MID26086310 Institution C Hispanic/Latinx Female   <NA>
#>  5: MID26000057 Institution C           White Female   <NA>
#>  6: MID26561940 Institution J           White   Male   <NA>
#>  7: MID25846316 Institution M           White   Male 143501
#>  8: MID25847220 Institution M           White   Male 143501
#>  9: MID25848589 Institution M           White   Male 143501
#> 10: MID25852023 Institution M           White   Male 143501
#> 11: MID25828870 Institution M           White   Male 149999

# Overwrites institution if present in dframe
DT <- toy_student[, .(mcid, institution, race, sex)]
condition_fye(dframe = DT, midfield_term = toy_term)

#>            mcid   institution            race    sex   cip6
#>  1: MID26060301 Institution C           Asian Female   <NA>
#>  2: MID25995980 Institution C Hispanic/Latinx Female   <NA>
#>  3: MID25997636 Institution C Hispanic/Latinx Female   <NA>
#>  4: MID26086310 Institution C Hispanic/Latinx Female   <NA>
#>  5: MID26000057 Institution C           White Female   <NA>
#>  6: MID26561940 Institution J           White   Male   <NA>
#>  7: MID25846316 Institution M           White   Male 143501
#>  8: MID25847220 Institution M           White   Male 143501
#>  9: MID25848589 Institution M           White   Male 143501
#> 10: MID25852023 Institution M           White   Male 143501
#> 11: MID25828870 Institution M           White   Male 149999

# Other columns, if any, are dropped
colnames(toy_student)

#> [1] "mcid"           "institution"    "transfer"       "hours_transfer"
#> [5] "race"           "sex"           

colnames(condition_fye(toy_student, toy_term))

#> [1] "mcid"        "institution" "race"        "sex"         "cip6"       

# Optional argument permits multiple CIP codes for FYE
condition_fye(DT, toy_term, fye_codes = c("140101", "140102"))

#>            mcid   institution            race    sex   cip6
#>  1: MID26060301 Institution C           Asian Female   <NA>
#>  2: MID25995980 Institution C Hispanic/Latinx Female   <NA>
#>  3: MID25997636 Institution C Hispanic/Latinx Female   <NA>
#>  4: MID26086310 Institution C Hispanic/Latinx Female   <NA>
#>  5: MID26000057 Institution C           White Female   <NA>
#>  6: MID26561940 Institution J           White   Male   <NA>
#>  7: MID25846316 Institution M           White   Male 143501
#>  8: MID25847220 Institution M           White   Male 143501
#>  9: MID25848589 Institution M           White   Male 143501
#> 10: MID25852023 Institution M           White   Male 143501
#> 11: MID25828870 Institution M           White   Male 149999