Condition multiway data for graphing

Transform a data frame such that two categorical variables are factors with levels ordered by medians of the quantitative variable. Rows and columns of the data frame are otherwise unaffected. The primary goal is to structure the data for display in a Cleveland "multiway dot plot," where the ordering of the levels and panels are crucial to the perception of effects.

condition_multiway(dframe, ..., details = NULL)

Arguments

dframe
Data frame of multiway data.

...
Not used, forces later arguments to be used by name.

details
Optional flag to add columns reporting the medians on which the order of the levels is based, default FALSE.

Value

A data.table with the following properties:

Rows are not modified.

Quantitative column is not modified.

Categorical columns are factors with levels ordered by median quantitative values with an option to add columns of group medians.

Grouping structures are not preserved.

Details

In multiway data---as defined by Cleveland (1993) ---there is a single quantitative value (or response) for every combination of levels of two categorical variables. condition_multiway() converts the columns of categorical variables to factors and orders the factor levels by increasing medians of the quantitative response variable.

Note that "multiway" in our context refers to the data structure and graph design defined by Cleveland, not to the methods of analysis described by Kroonenberg (2008) .

In the multiway dot plot, there are panels, the individual dot plots of the display, and there are levels, the rows of each panel. One category is encoded by the panels; the other by the rows. All panels have the same quantitative scale on the x-axis and the same organization of category levels on the y-axis. Panels and rows are ordered so that medians increase in "graph order", that is, from left to right and from bottom to top. Hence the importance of the medians reported by the details argument.

In our implementation, the dframe argument must have three columns only: two categorical columns (type character or factor) and one quantitative column (type numeric).

References

Cleveland WS (1993). Visualizing Data. Hobart Press, Summit, NJ.

Kroonenberg PM (2008). Applied Multiway Data Analysis. Wiley, Hoboken, NJ.

See also

Examples

# Construct example data frame
catg1 <- rep(c("urban", "rural", "suburb", "village"), each = 2)
catg2 <- rep(c("men", "women"), times = 4)
value <- c(0.22, 0.14, 0.43, 0.58, 0.81, 0.46, 0.15, 0.20)
dframe <- data.frame(catg1, catg2, value)
dframe

#>     catg1 catg2 value
#> 1   urban   men  0.22
#> 2   urban women  0.14
#> 3   rural   men  0.43
#> 4   rural women  0.58
#> 5  suburb   men  0.81
#> 6  suburb women  0.46
#> 7 village   men  0.15
#> 8 village women  0.20

# Class before conditioning
class(dframe$catg1)

#> [1] "character"

class(dframe$catg2)

#> [1] "character"

# Class and levels after conditioning
mw1 <- condition_multiway(dframe)
class(mw1$catg1)

#> [1] "factor"

levels(mw1$catg1)

#> [1] "village" "urban"   "rural"   "suburb" 

class(mw1$catg2)

#> [1] "factor"

levels(mw1$catg2)

#> [1] "men"   "women"

# Existing factors (if any) are re-ordered
mw2 <- dframe
mw2$catg1 <- factor(mw2$catg1,
                       levels = c("rural", "suburb", "urban", "village"))
levels(mw2$catg1) # levels before conditioning

#> [1] "rural"   "suburb"  "urban"   "village"

mw2 <- condition_multiway(mw2)
levels(mw2$catg1) # levels after conditioning

#> [1] "village" "urban"   "rural"   "suburb" 

# Report median values
mw <- condition_multiway(dframe, details = TRUE)
mw

#>      catg1 catg2 med_catg1 med_catg2 value
#> 1:   urban   men     0.180     0.325  0.22
#> 2:   urban women     0.180     0.330  0.14
#> 3:   rural   men     0.505     0.325  0.43
#> 4:   rural women     0.505     0.330  0.58
#> 5:  suburb   men     0.635     0.325  0.81
#> 6:  suburb women     0.635     0.330  0.46
#> 7: village   men     0.175     0.325  0.15
#> 8: village women     0.175     0.330  0.20