Deprecated functions in midfieldr

These functions were deprecated in midfieldr 1.0.4.

select_required()

is deprecated in favor of select_basic_cols(). The new functionality is similar but with exact matching to the default column names plus preserving keys and data frame class.

Deprecated in favor of select_basic_cols(dframe, patternv).

Usage

select_required(midfield_x, select_add = NULL)

Arguments

midfield_x

Data frame

select_add

Character vector