Add a column to evaluate record length for fair assessment

A logical variable is added to a data frame indicating whether the data available from an institution have sufficient span to fairly assess a student's record.

eval_fair_record(dframe, ..., record = NULL, details = NULL)

Arguments

dframe
data frame

...
not used, forces later arguments to be used by name

record
data frame of term attributes, default midfieldterms

details
logical scalar to add columns reporting information on which the evaluation is based, default FALSE

Value

Data frame with the following properties:

Rows are not modified

Column fair_record is added, column inst_limit is added optionally

Data frame attributes tbl or data.table are preserved

Grouping structures are not preserved

Details

Program completion is typically assessed over a given span of years after matriculation. A student matriculating too near the last term in the available data should be excluded from analysis because the data have insufficient span to fairly assess the student's record.

The data frame argument must include the timely_limit column obtained using the add_timely_limit() function. Assessment is considered fair if the student's timely completion term is no later than the last term in their institution's data.

If the result in the fair_record column is TRUE, then then student should be included in the research. If FALSE, the student should be excluded before calculating any persistence metric involving program completion (graduation). The function performs no subsetting.

If details is TRUE, additional column(s) that support the finding are returned as well. Here the extra column inst_limit, the latest term reported by the institution in the available data.

See also

Examples

# TBD