Example CIP excerpt

Excerpt of CIP data for a case study.

exa_cip

Format

A data frame (tibble) with 12 observations and 6 variables. The variables are characters.

cip6

6-digit program code.

cip6name

6-digit program name.

cip4

4-digit program code.

cip4name

4-digit program name.

cip2

2-digit program code.

cip2name

2-digit program name.

Source

midfielddata package

Details

An excerpt from the cip data set for Civil, Electrical, Industrial, and Mechanical Engineering. The data are an intermediate result in a case study developed in the midfieldr vignettes.

See also

Examples

exa_cip

#> # A tibble: 12 x 6
#>    cip2  cip2name  cip4  cip4name                cip6   cip6name                
#>    <chr> <chr>     <chr> <chr>                   <chr>  <chr>                   
#>  1 14    Engineer~ 1408  Civil Engineering       140801 Civil Engineering, Gene~
#>  2 14    Engineer~ 1408  Civil Engineering       140802 Geotechnical Engineering
#>  3 14    Engineer~ 1408  Civil Engineering       140803 Structural Engineering  
#>  4 14    Engineer~ 1408  Civil Engineering       140804 Transportation and High~
#>  5 14    Engineer~ 1408  Civil Engineering       140805 Water Resources Enginee~
#>  6 14    Engineer~ 1408  Civil Engineering       140899 Civil Engineering, Other
#>  7 14    Engineer~ 1410  Electrical, Electronic~ 141001 Electrical, Electronics~
#>  8 14    Engineer~ 1410  Electrical, Electronic~ 141003 Laser and Optical Engin~
#>  9 14    Engineer~ 1410  Electrical, Electronic~ 141004 Telecommunications Engi~
#> 10 14    Engineer~ 1410  Electrical, Electronic~ 141099 Electrical, Electronics~
#> 11 14    Engineer~ 1419  Mechanical Engineering  141901 Mechanical Engineering  
#> 12 14    Engineer~ 1435  Industrial Engineering  143501 Industrial Engineering  

cip6_select(data = exa_cip, program = "cip4name")