Subset rows by program CIP

Retain rows that match program CIPs. Default behavior is to retain all columns. Use keep_col to subset columns. Use unique_row to remove duplicate rows after subsetting.

filter_by_cip(
  data,
  keep_cip,
  ...,
  keep_col = NULL,
  unique_row = NULL,
  first_degree = NULL
)

Arguments

data
data frame

keep_cip
character vector of 6-digit CIP codes

...
not used, force later arguments to be used by name

keep_col
character vector of column names, default all columns

unique_row
logical, remove duplicate rows, default FALSE

first_degree
logical, retain first degree term only, default FALSE

Value

data frame with the following properties:

Rows with CIP matching an element of keep_cip

All columns if keep_col NULL, otherwise columns named in keep_col

Grouping structures are not preserved

Data frame extensions such as tbl or data.table are preserved

Details

The data argument---a midfielddata data set or equivalent---must include a cip6 column.

If applied to midfielddegrees, setting first_degree to TRUE retains the first degree term only (by id), dropping rows with degrees earned in subsequent terms. However, multiple degrees in the first degree term are always retained.

See also

Examples

# placeholder