Subset rows of a data frame by text

Retain rows that contain matching or partially matching character strings. Not case-sensitive. Default behavior is to retain all columns. Use keep_col to subset columns. Use unique_row to remove duplicate rows after subsetting.

filter_by_text(
  data,
  keep_text = NULL,
  ...,
  drop_text = NULL,
  keep_col = NULL,
  unique_row = NULL
)

Arguments

data
data frame of character columns

keep_text
character vector of search patterns for retaining rows

...
not used, force later arguments to be used by name

drop_text
character vector of search patterns for dropping rows

keep_col
character vector of column names, default all columns

unique_row
logical, remove duplicate rows, default FALSE

Value

data frame with the following properties:

Rows are a subset of the input in the same order

All columns if keep_col NULL, otherwise columns named in keep_col

Grouping structures are not preserved

Data frame extensions such as tbl or data.table are preserved

Details

The data argument is typically the midfieldr cip data set, though any data frame with all character-string columns can be searched and subset.

See also

Examples

filter_by_text(cip, keep_text = c("^1407", "^1410"))

#>    cip2    cip2name cip4                                               cip4name
#> 1:   14 Engineering 1407                                   Chemical Engineering
#> 2:   14 Engineering 1407                                   Chemical Engineering
#> 3:   14 Engineering 1407                                   Chemical Engineering
#> 4:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 5:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 6:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 7:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#>      cip6                                                      cip6name
#> 1: 140701                                          Chemical Engineering
#> 2: 140702                         Chemical and Biomolecular Engineering
#> 3: 140799                                   Chemical Engineering, Other
#> 4: 141001        Electrical, Electronics and Communications Engineering
#> 5: 141003                                 Laser and Optical Engineering
#> 6: 141004                                Telecommunications Engineering
#> 7: 141099 Electrical, Electronics and Communications Engineering, Other

filter_by_text(cip, keep_text = "civil engineering", drop_text = "technology")

#>    cip2    cip2name cip4          cip4name   cip6
#> 1:   14 Engineering 1408 Civil Engineering 140801
#> 2:   14 Engineering 1408 Civil Engineering 140802
#> 3:   14 Engineering 1408 Civil Engineering 140803
#> 4:   14 Engineering 1408 Civil Engineering 140804
#> 5:   14 Engineering 1408 Civil Engineering 140805
#> 6:   14 Engineering 1408 Civil Engineering 140899
#>                                  cip6name
#> 1:             Civil Engineering, General
#> 2:               Geotechnical Engineering
#> 3:                 Structural Engineering
#> 4: Transportation and Highway Engineering
#> 5:            Water Resources Engineering
#> 6:               Civil Engineering, Other

filter_by_text(cip, keep_text = "History") %>%
  filter_by_text(keep_text = "American")

#>    cip2 cip2name cip4 cip4name   cip6                         cip6name
#> 1:   54  History 5401  History 540102 American History (United States)