Filter CIP program data

Filter CIP program data by type and series. Type is the CIP taxonomy level: 2-digit, 4-digit, or 6-digit. Series is the CIP code.

filter_cip(type = 2, series = NULL, data = NULL)

Arguments

type
An integer (2, 4, or 6) representing the CIP taxonomy level

series
The first few digits of a CIP code or vector of codes. A character vector, or coercible to one. Can also be one of the predefined vectors of CIP codes for Engineering, Humanities, Mathematics, Sciences, or the Social Sciences (series_engr, series_hum, series_math, series_sci, series_socsci).

data
CIP data frame. The default is the midfieldr cip dataset with two variables, code and name.

Value

A subset of the cip dataset filtered by type and series

Examples

# Extract all 2-digit CIP data
filter_cip()

#> # A tibble: 48 x 2
#>     code                                                     name
#>    <chr>                                                    <chr>
#>  1    01 Agriculture/Agricultural Operations and Related Sciences
#>  2    03                       Natural Resources and Conservation
#>  3    04                        Architecture and Related Services
#>  4    05                  Area/Ethnic/Cultural and Gender Studies
#>  5    09           Communications/Journalism and Related Programs
#>  6    10         Communications Technologies and Support Services
#>  7    11   Computer and information Sciences and Support Services
#>  8    12                           Personal and Culinary Services
#>  9    13                                                Education
#> 10    14                                              Engineering
#> # ... with 38 more rows

# Extract all 4-digit CIP data
filter_cip(type = 4)

#> # A tibble: 411 x 2
#>     code                                                 name
#>    <chr>                                                <chr>
#>  1  0100                                  Agriculture/General
#>  2  0101                 Agricultural Business and Management
#>  3  0102                           Agricultural Mechanization
#>  4  0103                   Agricultural Production Operations
#>  5  0104            Agricultural and Food Products Processing
#>  6  0105            Agricultural and Domestic Animal Services
#>  7  0106 Applied Horticulture/Horticultural Business Services
#>  8  0107                            International Agriculture
#>  9  0108                         Agricultural Public Services
#> 10  0109                                      Animal Sciences
#> # ... with 401 more rows

# Extract all 4-digit CIP data in Engineering
filter_cip(type = 4, series = "14")

#> # A tibble: 40 x 2
#>     code                                                   name
#>    <chr>                                                  <chr>
#>  1  1401                                    Engineering/General
#>  2  1402   Aerospace/Aeronautical and Astronautical Engineering
#>  3  1403 Agricultural/Biological Engineering and Bioengineering
#>  4  1404                              Architectural Engineering
#>  5  1405                         Biomedical/Medical Engineering
#>  6  1406                       Ceramic Sciences and Engineering
#>  7  1407                                   Chemical Engineering
#>  8  1408                                      Civil Engineering
#>  9  1409                           Computer Engineering/General
#> 10  1410  Electrical/Electronics and Communications Engineering
#> # ... with 30 more rows

# Use the predefined Engineering series
filter_cip(type = 4, series = series_engr)

#> # A tibble: 40 x 2
#>     code                                                   name
#>    <chr>                                                  <chr>
#>  1  1401                                    Engineering/General
#>  2  1402   Aerospace/Aeronautical and Astronautical Engineering
#>  3  1403 Agricultural/Biological Engineering and Bioengineering
#>  4  1404                              Architectural Engineering
#>  5  1405                         Biomedical/Medical Engineering
#>  6  1406                       Ceramic Sciences and Engineering
#>  7  1407                                   Chemical Engineering
#>  8  1408                                      Civil Engineering
#>  9  1409                           Computer Engineering/General
#> 10  1410  Electrical/Electronics and Communications Engineering
#> # ... with 30 more rows