Subset rows by text

Subset a data frame, retaining rows that match or partially match character strings. All columns retained unless specified by keep_col. The function uses grepl(), therefore non-character columns that can be coerced to character are also searched for matches.

filter_text(
  data,
  keep_text = NULL,
  ...,
  drop_text = NULL,
  keep_col = NULL,
  unique_row = NULL
)

Arguments

data
data frame to be subset

keep_text
character vector of search patterns for retaining rows

...
not used, force later arguments to be used by name

drop_text
character vector of search patterns for dropping rows

keep_col
character vector of column names, default all columns

unique_row
logical, remove duplicate rows, default TRUE

Value

Data frame with the following properties:

Rows matching elements of keep_text

Columns specified by keep_col

Data frame extensions such as tbl or data.table are preserved

Grouping structures are not preserved

Examples

# filter using keywords
filter_text(cip, keep_text = "engineering")

#>      cip2                                         cip2name cip4
#>   1:   14                                      Engineering 1401
#>   2:   14                                      Engineering 1401
#>   3:   14                                      Engineering 1402
#>   4:   14                                      Engineering 1403
#>   5:   14                                      Engineering 1404
#>  ---                                                           
#> 117:   15                           Engineering Technology 1516
#> 118:   15                           Engineering Technology 1599
#> 119:   29                            Military Technologies 2903
#> 120:   29                            Military Technologies 2903
#> 121:   51 Health Professions and Related Clinical Sciences 5123
#>                                                     cip4name   cip6
#>   1:                                    Engineering, General 140101
#>   2:                                    Engineering, General 140102
#>   3:   Aerospace, Aeronautical and Astronautical Engineering 140201
#>   4: Agricultural, Biological Engineering and Bioengineering 140301
#>   5:                               Architectural Engineering 140401
#>  ---                                                               
#> 117:                                          Nanotechnology 151601
#> 118:    Engineering-Related Technologies, Technicians, Other 159999
#> 119:                               Military Applied Sciences 290301
#> 120:                               Military Applied Sciences 290303
#> 121:              Rehabilitation and Therapeutic Professions 512312
#>                                                              cip6name
#>   1:                                             Engineering, General
#>   2:                                                  Pre-Engineering
#>   3:     Aerospace, Aeronautical and Astronautical, Space Engineering
#>   4:          Agricultural, Biological Engineering and Bioengineering
#>   5:                                        Architectural Engineering
#>  ---                                                                 
#> 117:                                                   Nanotechnology
#> 118:             Engineering Related Technologies, Technicians, Other
#> 119:                                       Combat Systems Engineering
#> 120:                                            Engineering Acoustics
#> 121: Assistive, Augmentative Technology and Rehabiliation Engineering

# drop_text argument must be named
filter_text(cip, "civil engineering", drop_text = "technology")

#>    cip2    cip2name cip4          cip4name   cip6
#> 1:   14 Engineering 1408 Civil Engineering 140801
#> 2:   14 Engineering 1408 Civil Engineering 140802
#> 3:   14 Engineering 1408 Civil Engineering 140803
#> 4:   14 Engineering 1408 Civil Engineering 140804
#> 5:   14 Engineering 1408 Civil Engineering 140805
#> 6:   14 Engineering 1408 Civil Engineering 140899
#>                                  cip6name
#> 1:             Civil Engineering, General
#> 2:               Geotechnical Engineering
#> 3:                 Structural Engineering
#> 4: Transportation and Highway Engineering
#> 5:            Water Resources Engineering
#> 6:               Civil Engineering, Other

# filter using numerical codes
filter_text(cip, keep_text = c("050125", "160501"))

#>    cip2                                            cip2name cip4
#> 1:   05 Area, Ethnic, Cultural and Gender and Group Studies 0501
#> 2:   16      Foreign Languages, Literatures and Linguistics 1605
#>                                       cip4name   cip6
#> 1:                                Area Studies 050125
#> 2: Germanic Languages, Literatures Linguistics 160501
#>                          cip6name
#> 1:                 German Studies
#> 2: German Language and Literature

# filter using regular expressions
filter_text(cip, keep_text = "^54")

#>    cip2 cip2name cip4 cip4name   cip6
#> 1:   54  History 5401  History 540101
#> 2:   54  History 5401  History 540102
#> 3:   54  History 5401  History 540103
#> 4:   54  History 5401  History 540104
#> 5:   54  History 5401  History 540105
#> 6:   54  History 5401  History 540106
#> 7:   54  History 5401  History 540107
#> 8:   54  History 5401  History 540108
#> 9:   54  History 5401  History 540199
#>                                               cip6name
#> 1:                                    History, General
#> 2:                    American History (United States)
#> 3:                                    European History
#> 4:    History and Philosophy of Science and Technology
#> 5: Public, Applied History and Archival Administration
#> 6:                                       Asian History
#> 7:                                    Canadian History
#> 8:                                    Military History
#> 9:                                      History, Other

filter_text(cip, keep_text = c("^1407", "^1408"))

#>    cip2    cip2name cip4             cip4name   cip6
#> 1:   14 Engineering 1407 Chemical Engineering 140701
#> 2:   14 Engineering 1407 Chemical Engineering 140702
#> 3:   14 Engineering 1407 Chemical Engineering 140799
#> 4:   14 Engineering 1408    Civil Engineering 140801
#> 5:   14 Engineering 1408    Civil Engineering 140802
#> 6:   14 Engineering 1408    Civil Engineering 140803
#> 7:   14 Engineering 1408    Civil Engineering 140804
#> 8:   14 Engineering 1408    Civil Engineering 140805
#> 9:   14 Engineering 1408    Civil Engineering 140899
#>                                  cip6name
#> 1:                   Chemical Engineering
#> 2:  Chemical and Biomolecular Engineering
#> 3:            Chemical Engineering, Other
#> 4:             Civil Engineering, General
#> 5:               Geotechnical Engineering
#> 6:                 Structural Engineering
#> 7: Transportation and Highway Engineering
#> 8:            Water Resources Engineering
#> 9:               Civil Engineering, Other

# select columns
filter_text(cip,
            keep_text = "^54",
            keep_col = c("cip4", "cip4name"))

#>    cip4 cip4name
#> 1: 5401  History

if (FALSE) {
# unsuccessful terms produce a message
filter_text(cip, c("050125", "111111", "160501", "Bogus", "^55"))
}