Predicted starting programs for FYE students

Predicted starting programs for all First-Year Engineering (FYE) students in the midfielddata data tables. Keyed by student ID. Contains 6-digit CIP codes that are substituted for the First-Year-Engineering (FYE) CIP code when an analysis requires starting programs that are degree-granting. Suitable for use with midfielddata practice data only.

Usage

fye_predicted_start

Format

data.table with 5033 rows and 2 columns keyed by student ID. The variables are:

mcid

Character, de-identified student ID.

start

Character, 6-digit CIP code of the predicted starting program.

Details

Starting programs in fye_predicted_start are the engineering programs we predict that students would have declared had they not been required to enroll in FYE. The prediction is based on their first post-FYE program and a multivariate imputation suitable for multiple categorical variables using the mice package. The predictor variables are institution, race, and sex. The predicted variable is the 6-digit CIP code of a degree-granting engineering program at their institution.

fye_predicted_start holds only for the practice data in midfielddata---these values cannot be commingled with the research database available to MIDFIELD partners.

See also