Gather students starting in programs

Filters a data frame of student matriculation information (ID and starting major) to keep those starting in specified programs.

gather_start(.data, cip_group, transfer = FALSE)

Arguments

.data
A data frame of student matriculation information with two required character variables: student id and the cip6 code for their starting major. Default is midfieldstudents.

cip_group
A data frame with two required character variables: cip6 or start (6-digit CIP codes) and program (program labels).

transfer
Logical to include (TRUE) or exclude (FALSE) transfer students. Default is FALSE.

Value

A data frame with variables id (unique MIDFIELD student ID), start (6-digit CIP code), and program (a program label).

Details

gather_start() uses dplyr::inner_join() to combine the rows and columns of its two arguments, both data frames. All variables in the two arguments except id, cip6 or start and program are quietly dropped.

gather_start() excludes transfer students unless you set transfer = TRUE. If you are using the result to compute graduation rates per the IPEDS definition, transfer students should be omitted.

Examples

cip_group <- cip_filter(cip, series = "^54")
cip_group <- cip_label(cip_group, program = "cip2name")
starters <- gather_start(cip_group = cip_group)
head(starters)

#> # A tibble: 6 x 2
#>   id          start 
#>   <chr>       <chr> 
#> 1 MID25783780 540101
#> 2 MID25784804 540101
#> 3 MID25785329 540101
#> 4 MID25785903 540101
#> 5 MID25787024 540101
#> 6 MID25791292 540101