Subset CIP data frame

Subset the rows of a data frame by matching character patterns. The data are CIP attributes with 2-, 4-, and 6-digit codes and names in character columns (cip or equivalent). Rows are subset by matching or partially matching patterns.

get_cip(data = NULL, keep_any = NULL, ..., drop_any = NULL)

Arguments

data
data frame of CIP codes and names

keep_any
character vector of search patterns for retaining rows

...
not used for values. Forces optional arguments to be usable by name only.

drop_any
(optional) character vector of search patterns for dropping rows

Value

data.frame with the following properties:

Rows are a subset of the input in the same order

Columns are not modified

Grouping structures, if any, are not preserved

Data frame extension attributes, e.g., tibble, are not preserved

Details

The default data argument is cip, accessed by name or by leaving the data argument vacant. Any other data argument should be a subset of cip or equivalent.

get_cip() wraps the filter_char_frame operation. The search is not case-sensitive.

See also

Examples

get_cip(cip, keep_any = c("^1407", "^1410"))

#>   cip2    cip2name cip4                                               cip4name
#> 1   14 Engineering 1407                                   Chemical Engineering
#> 2   14 Engineering 1407                                   Chemical Engineering
#> 3   14 Engineering 1407                                   Chemical Engineering
#> 4   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 5   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 6   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 7   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#>     cip6                                                      cip6name
#> 1 140701                                          Chemical Engineering
#> 2 140702                         Chemical and Biomolecular Engineering
#> 3 140799                                   Chemical Engineering, Other
#> 4 141001        Electrical, Electronics and Communications Engineering
#> 5 141003                                 Laser and Optical Engineering
#> 6 141004                                Telecommunications Engineering
#> 7 141099 Electrical, Electronics and Communications Engineering, Other

get_cip(cip, keep_any = "civil engineering", drop_any = "technology")

#>   cip2    cip2name cip4          cip4name   cip6
#> 1   14 Engineering 1408 Civil Engineering 140801
#> 2   14 Engineering 1408 Civil Engineering 140802
#> 3   14 Engineering 1408 Civil Engineering 140803
#> 4   14 Engineering 1408 Civil Engineering 140804
#> 5   14 Engineering 1408 Civil Engineering 140805
#> 6   14 Engineering 1408 Civil Engineering 140899
#>                                 cip6name
#> 1             Civil Engineering, General
#> 2               Geotechnical Engineering
#> 3                 Structural Engineering
#> 4 Transportation and Highway Engineering
#> 5            Water Resources Engineering
#> 6               Civil Engineering, Other

get_cip(cip, keep_any = "History") %>%
  get_cip(keep_any = "American")

#>   cip2 cip2name cip4 cip4name   cip6                         cip6name
#> 1   54  History 5401  History 540102 American History (United States)