Get limits on feasible completion by institution

Determine the data limit and matriculation limit for program completion feasibility by institution. The data are term attributes keyed by institution and term (midfieldterms or equivalent). Rows are subset by the latest term in the data set (the data limit) by institution. Results are the institutions, matriculation limits, and data limits.

get_institution_limits(data = NULL, span = NULL)

Arguments

data
data frame of term attributes

span
number of years for feasible completion

Value

data.frame with the following properties:

One row per institution

Columns institution, matric_limit, and data_limit

Grouping structures, if any, are not preserved

Data frame extensions tbl or data.table are preserved

Details

The default data argument is midfieldterms, accessed by name or by leaving the data argument vacant. Any other data argument should be a subset of midfieldterms or equivalent. Character columns institution and term are required by name.

See also

Examples

# placeholder