Get degree status of students by ID

Subset the rows of a data frame by student ID. The data are student degree attributes keyed by student ID (midfielddegrees or equivalent). Results are student ID, institution, and degree.

get_status_degrees(data = NULL, keep_id = NULL)

Arguments

data
data frame of degree attributes

keep_id
character vector of student IDs

Value

data.frame with the following properties:

Rows for which student ID is in keep_id

Columns id, institution, and degree

Grouping structures, if any, are not preserved

Data frame extensions tbl or data.table are preserved

Details

The default data argument is midfielddegrees, accessed by name or by leaving the data argument vacant. Any other data argument should be a subset of midfielddegrees or equivalent. Character columns id, institution, and degree are required by name.

If the keep_id argument is NULL, all IDs in data are used by default.

See also

Examples

music_codes <- c("500901", "500903", "500913")
students <- get_enrollees(codes = music_codes)
students <- get_status_degrees(keep_id = students$id)
str(students)

#> Classes 'data.table' and 'data.frame':	745 obs. of  3 variables:
#>  $ id         : chr  "MID25853762" "MID25854314" "MID25855043" "MID25857363" ...
#>  $ institution: chr  "Institution B" "Institution B" "Institution B" "Institution B" ...
#>  $ degree     : chr  NA "Bachelor's Degree" "Bachelor's Degree" "Bachelor's Degree" ...
#>  - attr(*, ".internal.selfref")=<externalptr>