Gather graduates

Filter academic term data to find all students graduating from specified programs.

grad_filter(series, ..., reference = NULL, id = "id", cip6 = "cip6",
  term = "term_degree")

Arguments

series
atomic character vector specifying the 6-digit CIP codes of the programs to filter by

...
not used for values, forces later arguments to bind by name

reference
a reference data frame from which graduating student IDs, CIP codes, and terms are obtained, default midfielddegrees

id
character column name of the ID variable in reference

cip6
character column name of the CIP code variable in reference

term
character column name of term variable in reference

Value

Data frame with character variables for student ID and program CIP

Details

To use this function, the midfielddata package must be installed to provide midfielddegrees, the default reference data set.

The series argument is an atomic character vector of 6-digit CIP codes used to filter the reference data set and extract student IDs and terms.

The function returns a subset of reference with every unique combination of student ID and CIP of the program(s) in which they earned their first degree(s). Only the student ID and program CIP are returned; other variables in reference are quietly dropped.

Optional arguments. An alternate reference data set can be assigned via the reference argument. The alternate must have variables for student ID, 6-digit CIP, and terms in which degrees were earned with the same structure as in midfielddegrees. Use the id, cip6, and term arguments to reassign the variables names, if necessary, to match the variable names in the alternate.

See also

Examples

grad <- grad_filter(series = "540104")
grad

#> # A tibble: 24 x 2
#>    id          cip6  
#>    <chr>       <chr> 
#>  1 MID26360410 540104
#>  2 MID26361391 540104
#>  3 MID26362301 540104
#>  4 MID26362581 540104
#>  5 MID26365694 540104
#>  6 MID26365705 540104
#>  7 MID26366791 540104
#>  8 MID26367223 540104
#>  9 MID26368839 540104
#> 10 MID26383658 540104
#> # ... with 14 more rows