Graph functions

Provides a set of customized layers and arguments graphs created with for ggplot2.

expon_scale_x_log10()

midfield_theme()

darkBr

medBr

lightBr

paleBr

minBr

minBG

paleBG

lightBG

medBG

darkBG

darkPR

medPR

lightPR

palePR

minPR

minGn

paleGn

lightGn

medGn

darkGn

gray1

gray2

gray3

gray4

gray5

gray6

gray7

gray8

gray9

Format

An object of class character of length 1.

Value

Modifies an existing graph

Details

expon_scale_x_log10()

Applies the usual scale_x_log10() function with the scale marked powers of ten the exponent.

midfield_theme()

Applies theme_minimal() with some additional edits.

darkBr ... minBr

Brown color palette. Five hues named darkBr, medBr, lightBr, paleBr, minBr from the RColorBrewer "BrBG" palette.

darkBG ... minBG

Blue-green color palette. Five hues named darkBG, medBG, lightBG, paleBG, minBG from the RColorBrewer "BrBG" palette.

darkPR ... minPR

Purple color palette. Five hues named darkPR, medPR, lightPR, palePR, minPR from the RColorBrewer "PRGn" palette.

darkGn ... minBr

Green color palette. Five hues named darkGn, medGn, lightGn, paleGn, minGn from the RColorBrewer "PRGn" palette.

gray9 ... gray1

Gray color palette. Nine hues named gray9, gray8, ... , gray2, gray1 from the RColorBrewer "Greys" palette.