Label programs and extract 6-digit CIP columns

Extract two columns of a data frame and add a column of program labels. The data are a subset of midfieldr cip or equivalent. The result is a 3-column data frame with 6-digit CIP codes, 6-digit CIP names, and user-assigned program labels.

label_programs(data = NULL, label = NULL)

Arguments

data
data frame of CIP codes and names

label
scalar character string

Value

data.frame with the following properties:

Rows are not modified

The 6-digit code and name columns are not modified. Other columns are dropped and a program column is added.

Grouping structures, if any, are not preserved

Data frame extension attributes, e.g., tibble, are not preserved

Details

The typical data argument is an output of get_cip() with six columns: 2-digit codes and names, 4-digit codes and names, and 6-digit codes and names. label_programs() retains the two 6-digit columns and adds a program column for the labels. Equivalent to

data$program <- "my_chosen_label"
data[, c("cip6", "cip6name", "program")]

label_programs() facilitates selecting labels suitable for grouping, summarizing, and joining operations. One can assign a custom label or use the 2-, 4-, or 6-digit program names in CIP. The label argument has three options:

A string (in quotes), for example label = "Electrical engineering", places the string in every row of the new program column

If label = "cip2name", the labels are the 2-digit CIP program names. Argument "cip6name" produces similar results.

If label = NULL or "cip4name", the labels are the 4-digit CIP program names

See also

Examples

ece_cip <- get_cip(cip, keep_any = "^1410")
label_programs(ece_cip, label = "ECE")
label_programs(ece_cip, label = "cip2name")
label_programs(ece_cip, label = "cip6name")