Undergraduate Student Records For Longitudinal Research

A package for investigating a sample from the MIDFIELD database, a dataset compiled from registrar's student records from participating US universities. The sample data include demographic, term, course, and degree information for 165,000 undergraduate students from 1990 to 2016.

Details

The Multi-Institution Database for Investigating Engineering Longitudinal Development (MIDFIELD) was started with longitudinal studies of engineering students in mind (the "E" in MIDFIELD). However, the data comprise complete registrar's student records---any academic program at those institutions can be studied.

Data

Sample data from the MIDFIELD database are provided in 3 additional packages. The datasets are:

student

midfieldstudent package. Demographic data for 165,000 students. Each observation is a unique student.

term

midfieldterm package. Academic term data for 165,000 students. Each observation is one term for one student, yielding 1.1 M observations.

course

midfieldcourse package. Academic course data for 165,000 students. Each observation is one course for one student, yielding 5.4 M observations.

degree

midfieldstudent package. Graduation data for 90,000 students receiving degrees. Each observation is a unique student.

cip

midfieldr package. A dataset of program codes and names of academic fields of study. Each of the 1552 observations is one program at the CIP 6-digit level. ?cip for more information.