Named series of 6-digit CIP codes

Collections of 6-digit CIP codes denoting groups of programs used as the series argument in cip_filter().

bio_sci

engr

math_stat

other_stem

phys_sci

stem

Format

A character vector of 6-digit CIP codes

Source

NSF STEM programs: https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf

Details

While the 2-digit and 4-digit CIP series represent convenient groupings of programs, e.g., History (CIP series 54) or Civil Engineering (CIP series 1408), some groups of programs such as Science, Technology, Engineering, and Mathematics (STEM) programs are comprised of programs from different 2-digit series.

The midfieldr `named_series` datasets are collections of 6-digit codes for groups of programs from multiple 2-digit series, such as STEM programs, as well as for programs in a single 2-digit series such as Engineering.

The classification of STEM programs is from the US National Science Foundation (NSF).

bio_sci

Biological and Biomedical Sciences (CIP 26 series)

engr

Engineering (CIP 14 series)

math_stat

Mathematics and Statistics (CIP 27 series)

other_stem

All 6-digit CIP codes for STEM programs excluding series 14, 26, 27, and 40.

phys_sci

Physical Sciences (CIP 40 series)

stem

All 6-digit CIP codes for STEM programs. The concatenation of all codes in the named series bio_sci, engr, math_stat, other_stem, and phys_sci.

Examples

engr

#>  [1] "140101" "140102" "140201" "140301" "140401" "140501" "140601" "140701"
#>  [9] "140702" "140799" "140801" "140802" "140803" "140804" "140805" "140899"
#> [17] "140901" "140902" "140903" "140999" "141001" "141003" "141004" "141099"
#> [25] "141101" "141201" "141301" "141401" "141801" "141901" "142001" "142101"
#> [33] "142201" "142301" "142401" "142501" "142701" "142801" "143101" "143201"
#> [41] "143301" "143401" "143501" "143601" "143701" "143801" "143901" "144001"
#> [49] "144101" "144201" "144301" "144401" "144501" "149999"

cip_filter(engr)

#>    CIP2    CIP2Name CIP4                                               CIP4Name
#> 1    14 Engineering 1401                                    Engineering/General
#> 2    14 Engineering 1401                                    Engineering/General
#> 3    14 Engineering 1402   Aerospace/Aeronautical and Astronautical Engineering
#> 4    14 Engineering 1403 Agricultural/Biological Engineering and Bioengineering
#> 5    14 Engineering 1404                              Architectural Engineering
#> 6    14 Engineering 1405                         Biomedical/Medical Engineering
#> 7    14 Engineering 1406                       Ceramic Sciences and Engineering
#> 8    14 Engineering 1407                                   Chemical Engineering
#> 9    14 Engineering 1407                                   Chemical Engineering
#> 10   14 Engineering 1407                                   Chemical Engineering
#> 11   14 Engineering 1408                                      Civil Engineering
#> 12   14 Engineering 1408                                      Civil Engineering
#> 13   14 Engineering 1408                                      Civil Engineering
#> 14   14 Engineering 1408                                      Civil Engineering
#> 15   14 Engineering 1408                                      Civil Engineering
#> 16   14 Engineering 1408                                      Civil Engineering
#> 17   14 Engineering 1409                           Computer Engineering/General
#> 18   14 Engineering 1409                           Computer Engineering/General
#> 19   14 Engineering 1409                           Computer Engineering/General
#> 20   14 Engineering 1409                           Computer Engineering/General
#> 21   14 Engineering 1410  Electrical/Electronics and Communications Engineering
#> 22   14 Engineering 1410  Electrical/Electronics and Communications Engineering
#> 23   14 Engineering 1410  Electrical/Electronics and Communications Engineering
#> 24   14 Engineering 1410  Electrical/Electronics and Communications Engineering
#> 25   14 Engineering 1411                                  Engineering Mechanics
#> 26   14 Engineering 1412                                    Engineering Physics
#> 27   14 Engineering 1413                                    Engineering Science
#> 28   14 Engineering 1414         Environmental/Environmental Health Engineering
#> 29   14 Engineering 1418                                  Materials Engineering
#> 30   14 Engineering 1419                                 Mechanical Engineering
#> 31   14 Engineering 1420                              Metallurgical Engineering
#> 32   14 Engineering 1421                         Mining and Mineral Engineering
#> 33   14 Engineering 1422              Naval Architecture and Marine Engineering
#> 34   14 Engineering 1423                                    Nuclear Engineering
#> 35   14 Engineering 1424                                      Ocean Engineering
#> 36   14 Engineering 1425                                  Petroleum Engineering
#> 37   14 Engineering 1427                                    Systems Engineering
#> 38   14 Engineering 1428                       Textile Sciences and Engineering
#> 39   14 Engineering 1431                                      Materials Science
#> 40   14 Engineering 1432                           Polymer/Plastics Engineering
#> 41   14 Engineering 1433                               Construction Engineering
#> 42   14 Engineering 1434                                     Forest Engineering
#> 43   14 Engineering 1435                                 Industrial Engineering
#> 44   14 Engineering 1436                              Manufacturing Engineering
#> 45   14 Engineering 1437                                    Operations Research
#> 46   14 Engineering 1438                                  Surveying Engineering
#> 47   14 Engineering 1439                     Geological/Geophysical Engineering
#> 48   14 Engineering 1440                          Paper Science and Engineering
#> 49   14 Engineering 1441                          Electromechanical Engineering
#> 50   14 Engineering 1442       Mechatronics/Robotics and Automation Engineering
#> 51   14 Engineering 1443                                Biochemical Engineering
#> 52   14 Engineering 1444                                  Engineering Chemistry
#> 53   14 Engineering 1445                      Biological/Biosystems Engineering
#> 54   14 Engineering 1499                                      Engineering/Other
#>      CIP6                                                      CIP6Name
#> 1  140101                                         Engineering - General
#> 2  140102                                               Pre-Engineering
#> 3  140201          Aerospace/Aeronautical and Astronautical Engineering
#> 4  140301        Agricultural/Biological Engineering and Bioengineering
#> 5  140401                                     Architectural Engineering
#> 6  140501                                Biomedical/Medical Engineering
#> 7  140601                              Ceramic Sciences and Engineering
#> 8  140701                                          Chemical Engineering
#> 9  140702                         Chemical and Biomolecular Engineering
#> 10 140799                                  Chemical Engineering - Other
#> 11 140801                                   Civil Engineering - General
#> 12 140802                                      Geotechnical Engineering
#> 13 140803                                        Structural Engineering
#> 14 140804                        Transportation and Highway Engineering
#> 15 140805                                   Water Resources Engineering
#> 16 140899                                     Civil Engineering - Other
#> 17 140901                                Computer Engineering - General
#> 18 140902                                 Computer Hardware Engineering
#> 19 140903                                 Computer Software Engineering
#> 20 140999                                  Computer Engineering - Other
#> 21 141001         Electrical/Electronics and Communications Engineering
#> 22 141003                                 Laser and Optical Engineering
#> 23 141004                                Telecommunications Engineering
#> 24 141099 Electrical/Electronics and Communications Engineering - Other
#> 25 141101                                         Engineering Mechanics
#> 26 141201                                           Engineering Physics
#> 27 141301                                           Engineering Science
#> 28 141401                Environmental/Environmental Health Engineering
#> 29 141801                                         Materials Engineering
#> 30 141901                                        Mechanical Engineering
#> 31 142001                                     Metallurgical Engineering
#> 32 142101                                Mining and Mineral Engineering
#> 33 142201                     Naval Architecture and Marine Engineering
#> 34 142301                                           Nuclear Engineering
#> 35 142401                                             Ocean Engineering
#> 36 142501                                         Petroleum Engineering
#> 37 142701                                           Systems Engineering
#> 38 142801                              Textile Sciences and Engineering
#> 39 143101                                             Materials Science
#> 40 143201                                  Polymer/Plastics Engineering
#> 41 143301                                      Construction Engineering
#> 42 143401                                            Forest Engineering
#> 43 143501                                        Industrial Engineering
#> 44 143601                                     Manufacturing Engineering
#> 45 143701                                           Operations Research
#> 46 143801                                         Surveying Engineering
#> 47 143901                            Geological/Geophysical Engineering
#> 48 144001                                 Paper Science and Engineering
#> 49 144101                                 Electromechanical Engineering
#> 50 144201              Mechatronics/Robotics and Automation Engineering
#> 51 144301                                       Biochemical Engineering
#> 52 144401                                         Engineering Chemistry
#> 53 144501                             Biological/Biosystems Engineering
#> 54 149999                                           Engineering - Other