Prepare multiway data for graphing

Transform a data frame such that two categorical variables are factors with levels ordered by medians of the quantitative variable. Rows and columns of the data frame are otherwise unaffected.

prepare_multiway(data = NULL)

Arguments

data
data frame of multiway data

Value

data.frame with the following properties:

Rows are not modified

Quantitative column is not modified

Categorical columns are factors with levels ordered by median quantitative values

Grouping structures, if any, are not preserved

Data frame extensions tbl or data.table are preserved

Details

In multiway data, there is a single quantitative value (or response) for every combination of levels of two categorical variables. order_multiway() converts the columns of categorical variables to factors and orders the factor levels by increasing medians of the quantitative response variable.

The data argument, therefore, must have three columns only. One column is a quantitative variable (type numeric) and two columns are categorical variables (type character or factor).

See also

Examples

catg1 <- rep(c("urban", "rural", "suburb", "village"), each = 2)
catg2 <- rep(c("men", "women"), times = 4)
value <- c(0.22, 0.14, 0.43, 0.58, 0.81, 0.46, 0.15, 0.20)

# structure before
pre_mw <- data.frame(catg1, catg2, value)
str(pre_mw)

#> 'data.frame':	8 obs. of  3 variables:
#>  $ catg1: chr  "urban" "urban" "rural" "rural" ...
#>  $ catg2: chr  "men" "women" "men" "women" ...
#>  $ value: num  0.22 0.14 0.43 0.58 0.81 0.46 0.15 0.2

# structure after
mw <- prepare_multiway(pre_mw)
str(mw)

#> 'data.frame':	8 obs. of  3 variables:
#>  $ catg1: Factor w/ 4 levels "village","urban",..: 2 2 3 3 4 4 1 1
#>   ..- attr(*, "scores")= num [1:4(1d)] 0.505 0.635 0.18 0.175
#>   .. ..- attr(*, "dimnames")=List of 1
#>   .. .. ..$ : chr [1:4] "rural" "suburb" "urban" "village"
#>  $ catg2: Factor w/ 2 levels "men","women": 1 2 1 2 1 2 1 2
#>   ..- attr(*, "scores")= num [1:2(1d)] 0.325 0.33
#>   .. ..- attr(*, "dimnames")=List of 1
#>   .. .. ..$ : chr [1:2] "men" "women"
#>  $ value: num  0.22 0.14 0.43 0.58 0.81 0.46 0.15 0.2