Case CIP codes

Representative case study. A subset of the cip data set containing CIP codes and names for Civil, Electrical, Industrial, and Mechanical Engineering programs in midfielddata.

rep_cip

Format

data.table with 12 rows and 6 columns keyed by the 6-digit CIP code. The variables are:

cip6

character 6-digit code representing "specific instructional programs" (US National Center for Education Statistics)

cip6name

character program name at the 6-digit level

cip4

character 4-digit code (the first 4 digits of cip6) representing "intermediate groupings of programs that have comparable content and objectives"

cip4name

character program name at the 4-digit level

cip2

character 2-digit code (the first 2 digits of cip6) representing "the most general groupings of related programs"

cip2name

character program name at the 2-digit level

See also