Student ID data sets

Representative case study. Character vectors of student IDs at different points in a case study of Civil, Electrical, Industrial, and Mechanical Engineering programs in midfielddata.

rep_ever

rep_grad

rep_ever_fc

Format

character vector. The names of the available data vectors are:

rep_ever

IDs of unique students ever enrolled in the example programs

rep_ever_fc

subset of rep_ever for whom program completion is feasible in 6 years

rep_grad

IDs of unique students graduating from the example programs

See also

Examples

library(midfielddata)

# ever enrolled
rows_we_want <- midfieldstudents$id %in% rep_ever
enrollees <- midfieldstudents[rows_we_want]
head(enrollees)

#>             id   institution   cip6 term_enter transfer hours_transfer  sex
#> 1: MID25783178 Institution M 14XXXX      19946        N             NA Male
#> 2: MID25783197 Institution M 14XXXX      19871        N             NA Male
#> 3: MID25783388 Institution M 14XXXX      19871        N             NA Male
#> 4: MID25783441 Institution M 14XXXX      19921        N             NA Male
#> 5: MID25783491 Institution M 14XXXX      19961        N             NA Male
#> 6: MID25783553 Institution M 14XXXX      20021        N             NA Male
#>     race age us_citizen home_zip high_school sat_math sat_verbal act_comp
#> 1: Black  17          Y    29169      411290      642        486       24
#> 2: White  18          Y    16105      392902      650        556       25
#> 3: White  18          Y    29670      411620      560        455       21
#> 4: White  18          Y    19425      391005      680        486       24
#> 5: White  18          Y    11784      331283      540        555       24
#> 6: White  17          Y    29466      411445      640        536       24

# graduates
rows_we_want <- midfieldstudents$id %in% rep_grad
graduates <- midfieldstudents[rows_we_want]
head(graduates)

#>             id   institution   cip6 term_enter transfer hours_transfer    sex
#> 1: MID25783178 Institution M 14XXXX      19946        N             NA   Male
#> 2: MID25783197 Institution M 14XXXX      19871        N             NA   Male
#> 3: MID25783441 Institution M 14XXXX      19921        N             NA   Male
#> 4: MID25783491 Institution M 14XXXX      19961        N             NA   Male
#> 5: MID25783912 Institution M 14XXXX      19971        N             NA   Male
#> 6: MID25784334 Institution M 14XXXX      19941        Y             NA Female
#>     race age us_citizen home_zip high_school sat_math sat_verbal act_comp
#> 1: Black  17          Y    29169      411290      642        486       24
#> 2: White  18          Y    16105      392902      650        556       25
#> 3: White  18          Y    19425      391005      680        486       24
#> 4: White  18          Y    11784      331283      540        555       24
#> 5: White  18          Y    29620      410000      570        685       26
#> 6: White  20          Y    29644        <NA>      540        505       20