Subset ID vector for feasible completion

Subset a vector of student IDs, retaining students for whom program completion is feasible within a designated time limit.

subset_feasible(
  id,
  ...,
  span = NULL,
  term_advance_max = NULL,
  data_students = NULL,
  data_terms = NULL,
  data_degrees = NULL
)

Arguments

id
character vector of student IDs

...
not used, force later arguments to be used by name.

span
numeric scalar, number of years to define feasibility, default is 6 years

term_advance_max
numeric scalar, maximum terms of transfer credit that advance the matriculation limit, default is 4 terms

data_students
data frame of student attributes

data_terms
data frame of term attributes

data_degrees
data frame of degree attributes

Value

character vector of student IDs, a subset of the input

Details

To count a non-graduating student in a persistence metric, completing their instructional program must have been feasible. Completion feasibility depends on the range of data from the institution, the student's matriculation term, and a designated time limit (typically 6 years). The relevant terminology is:

data range

The span of terms for which student-record data are available. Varies by institution.

completion feasibility

A student may complete their program within the limits of the data range.

matriculation limit

The last term a student can matriculate and feasibly complete their program within a designated time limit (typically 6 years). Transfer credits can advance the matriculation limit.

The subsetting condition is that a student has graduated, or if not, that their matriculation term is no later than the matriculation limit, adjusted for transfer credits if any, up to a maximum of 4 terms equivalent.

The function accesses three of the midfielddata data sets: midfieldstudents, midfieldterms, and midfielddegrees or their equivalents.

Examples

# sample character vector of student IDs
id_sample <- sample(midfielddata::midfieldstudents$id, 100)
str(id_sample)

#>  chr [1:100] "MID26309635" "MID26499656" "MID26058611" "MID25954512" ...

# subset the vector for feasible completion
id_feasible <- subset_feasible(id_sample)
str(id_feasible)

#>  chr [1:80] "MID25792714" "MID25812730" "MID25820903" "MID25846051" ...

# using a span > 6 years
id_feasible <- subset_feasible(id_sample, span = 8)
str(id_feasible)

#>  chr [1:75] "MID25792714" "MID25812730" "MID25820903" "MID25846051" ...

# allow more than 4 terms of transfer credit
id_feasible <- subset_feasible(id_sample, term_advance_max = 6)
str(id_feasible)

#>  chr [1:80] "MID25792714" "MID25812730" "MID25820903" "MID25846051" ...