Determine data sufficiency for every student

Add a column to a data frame of Student Unit Record (SUR) observations that labels each row for inclusion or exclusion based on data sufficiency near the upper and lower bounds of an institution's data range. Requires a MIDFIELD term data frame in the environment.

Usage

add_data_sufficiency(dframe, midfield_term = term)

Arguments

dframe

Data frame of student unit record (SUR) observations keyed by student ID. Required variables are mcid and timely_term. See also add_timely_term().

midfield_term

Data frame of SUR term observations keyed by student ID. Default is term. Required variables are mcid, institution, and term.

Value

A data.table with the following properties:

Rows are not modified.

Grouping structures are not preserved.

Columns listed below are added. Caution! An existing column with the same name as one of the added columns is silently overwritten. Other columns are not modified.

Columns added:

term_i

Character. Initial term of a student's longitudinal record, encoded YYYYT. Not overwritten if present in dframe.

lower_limit

Character. Initial term of an institution's data range, encoded YYYYT

upper_limit

Character. Final term of an institution's data range, encoded YYYYT

data_sufficiency

Character. Label each observation for inclusion or exclusion based on data sufficiency. Possible values are: include, indicating that available data are sufficient for estimating timely program completion; exclude-upper, indicating that data are insufficient at the upper limit of a data range; and exclude-lower, indicating that data are insufficient at the lower limit.

Details

The time span of MIDFIELD term data varies by institution, each having their own lower and upper bounds. For some student records, being at or near these bounds creates unavoidable ambiguity when trying to assess program completion. Such records must be identified and in most cases excluded to prevent false summary counts.

See also

Examples

# Start with IDs
# dframe <- toy_student[1:10, .(mcid)]
# 
# 
# # Add the timely completion term, required for data sufficiency 
# dframe <- add_timely_term(dframe, midfield_term = toy_term)
# 
# 
# # Add the data sufficiency variable 
# add_data_sufficiency(dframe, midfield_term = toy_term)
# 
# 
# # Data sufficiency column with details
# add_data_sufficiency(dframe, midfield_term = toy_term, details = TRUE)
# 
# 
# # If present, existing data_sufficiency column is overwritten
# # Using dframe from above,
# DT1 <- add_data_sufficiency(dframe, midfield_term = toy_term)
# DT2 <- add_data_sufficiency(DT1, midfield_term = toy_term)
# all.equal(DT1, DT2)