Add a column to evaluate data sufficiency

Add a column of logical values (TRUE/FALSE) to a data frame indicating whether the available data include a sufficient range of years to justify including a student in an analysis. Based on information in the MIDFIELD term data table or equivalent.

add_data_sufficiency(dframe, midfield_term, ..., details = NULL)

Arguments

dframe
Data frame with required variables institution and timely_term.

midfield_term
MIDFIELD term data table or equivalent with required variables institution and term.

...
Not used, forces later arguments to be used by name.

details
Optional flag to add columns reporting information on which the evaluation is based, default FALSE.

Value

A data.table with the following properties:

Rows are not modified.

Column data_sufficiency is added with an option to add column inst_limit.

Grouping structures are not preserved.

Details

Program completion is typically considered timely if it occurs within a specific span of years after admission. Students admitted too near the last term in the available data are generally excluded from a study because the data have insufficient range to fairly assess their records.

The input data frame dframe must include the timely_term column obtained using the add_timely_term() function. Students can be retained in a study if their estimated timely completion term is no later than the last term in their institution's data.

If the result in the data_sufficiency column is TRUE, then then student should be included in the research. If FALSE, the student should be excluded before calculating any persistence metric involving program completion (graduation). The function itself performs no subsetting.

If details is TRUE, additional column(s) that support the finding are returned as well. Here the extra column is inst_limit, the latest term reported by the institution in the available data.

Existing columns with the same names as the added columns are overwritten.

See also

Examples

# Start with IDs and add institution and timely term
dframe <- toy_student[1:10, .(mcid)]
dframe <- add_institution(dframe, midfield_term = toy_term)
dframe <- add_timely_term(dframe, midfield_term = toy_term)

# Data sufficiency column
add_data_sufficiency(dframe, midfield_term = toy_term)
#>            mcid   institution timely_term data_sufficiency
#>  1: MID25783939 Institution M       20053             TRUE
#>  2: MID25784402 Institution M       20103            FALSE
#>  3: MID25805538 Institution M       20091             TRUE
#>  4: MID25808099 Institution M       19923             TRUE
#>  5: MID25816437 Institution M       20113            FALSE
#>  6: MID25826223 Institution M       19913             TRUE
#>  7: MID25828870 Institution M       19933             TRUE
#>  8: MID25831839 Institution M       20123            FALSE
#>  9: MID25839453 Institution M       20111            FALSE
#> 10: MID25840802 Institution M       19951             TRUE

# Data sufficiency column with details
add_data_sufficiency(dframe, midfield_term = toy_term, details = TRUE)
#>            mcid   institution timely_term inst_limit data_sufficiency
#>  1: MID25783939 Institution M       20053      20093             TRUE
#>  2: MID25784402 Institution M       20103      20093            FALSE
#>  3: MID25805538 Institution M       20091      20093             TRUE
#>  4: MID25808099 Institution M       19923      20093             TRUE
#>  5: MID25816437 Institution M       20113      20093            FALSE
#>  6: MID25826223 Institution M       19913      20093             TRUE
#>  7: MID25828870 Institution M       19933      20093             TRUE
#>  8: MID25831839 Institution M       20123      20093            FALSE
#>  9: MID25839453 Institution M       20111      20093            FALSE
#> 10: MID25840802 Institution M       19951      20093             TRUE

# If present, existing data_sufficiency column is overwritten
# Using dframe from above,
DT1 <- add_data_sufficiency(dframe, midfield_term = toy_term)
DT2 <- add_data_sufficiency(DT1, midfield_term = toy_term)
all.equal(DT1, DT2)
#> [1] TRUE