Add a column to evaluate data sufficiency

A column is added to a data frame indicating whether the data include a sufficient number of years to justify including a student in an analysis of student records. The new column is a logical variable (TRUE/FALSE values).

add_data_sufficiency(dframe, ..., mdata = NULL, details = NULL)

Arguments

dframe
data frame with required variables institution and timely_term

...
not used, forces later arguments to be used by name

mdata
MIDFIELD term data, default midfielddata::term, with required variables institution and term

details
logical scalar to add columns reporting information on which the evaluation is based, default FALSE

Value

A data.table with the following properties:

Rows are not modified

Column data_sufficiency is added, column inst_limit is added optionally

Grouping structures are not preserved

Details

Program completion is typically assessed over a given span of years after admission. A student admitted too near the last term in the available data should be excluded from analysis because the data have insufficient span to fairly assess the student's record.

The data frame argument must include the timely_term column obtained using the add_timely_term() function. Assessment is considered fair if the student's timely completion term is no later than the last term in their institution's data.

If the result in the data_sufficiency column is TRUE, then then student should be included in the research. If FALSE, the student should be excluded before calculating any persistence metric involving program completion (graduation). The function performs no subsetting.

If details is TRUE, additional column(s) that support the finding are returned as well. Here the extra column inst_limit, the latest term reported by the institution in the available data.

Examples

# TBD