Add columns for race/ethnicity and sex

Add variables race and sex from student to a data frame, using mcid as the join-by variable.

add_race_sex(dframe, ..., midfield_table = NULL)

Arguments

dframe
data frame with required variable mcid

...
not used, forces later arguments to be used by name

midfield_table
MIDFIELD student table, with required variables mcid, race, and sex

Value

A data.table with the following properties:

Rows are not modified

Columns race and sex are added

Grouping structures are not preserved

Details

Existing variables with the same names, if any, are overwritten.

Examples

# TBD