Add columns for student race/ethnicity and sex

Add columns for students' self-reported race/ethnicity and sex using student ID as the join-by variable. Obtains the information from the MIDFIELD student data table or equivalent.

Usage

add_race_sex(dframe, midfield_student)

Arguments

dframe

Data frame with required variable mcid.

midfield_student

MIDFIELD student data table or equivalent with required variables mcid, race, and sex.

Value

A data.table with the following properties:

Rows are not modified.

Columns race and sex are added.

Grouping structures are not preserved.

Details

Existing columns with the same names as the added columns are overwritten.

See also

Examples

# Add race and sex to a data frame of graduates
dframe <- toy_degree[1:5, c("mcid", "cip6")]
add_race_sex(dframe, midfield_student = toy_student)
#>           mcid   cip6  race    sex
#> 1: MID25783939 010901 White Female
#> 2: MID25808099 521401 White Female
#> 3: MID25826223 131202 White Female
#> 4: MID25828870 140801 White   Male
#> 5: MID25840802 520201 White Female

# Add race and sex to a data frame from the term table
dframe <- toy_term[21:26, c("mcid", "institution", "level")]
add_race_sex(dframe, midfield_student = toy_student)
#>           mcid   institution        level  race    sex
#> 1: MID25845841 Institution M    03 Junior White Female
#> 2: MID25846316 Institution M  01 Freshman White   Male
#> 3: MID25846316 Institution M 02 Sophomore White   Male
#> 4: MID25846316 Institution M 02 Sophomore White   Male
#> 5: MID25846316 Institution M 02 Sophomore White   Male
#> 6: MID25846316 Institution M 02 Sophomore White   Male

# If present, existing race and sex columns are overwritten
# Using dframe from above,
DT1 <- add_race_sex(dframe, midfield_student = toy_student)
DT2 <- add_race_sex(DT1, midfield_student = toy_student)
all.equal(DT1, DT2)
#> [1] TRUE