Add columns for race/ethnicity and sex

Add variables race and sex from record to a data frame, using id is the join-by variable.

add_race_sex(dframe, ..., record = NULL)

Arguments

dframe
data frame

...
not used, forces later arguments to be used by name

record
data frame of student attributes, default midfieldstudents

Value

Data frame with the following properties:

Rows are not modified

Columns race and sex are added

Data frame attributes tbl or data.table are preserved

Grouping structures are not preserved

Details

Existing variables with the same names, if any, are overwritten.

Examples

# TBD