Add a column of timely completion terms

Add a column of academic term values to a data frame indicating the latest term by which a student can graduate and have it considered a timely completion. Student ID is the join-by variable; terms are encoded as character strings YYYYT. Obtains the information from the MIDFIELD term data table or equivalent.

add_timely_term(
  dframe,
  midfield_term,
  ...,
  details = NULL,
  span = NULL,
  sched_span = NULL
)

Arguments

dframe
Data frame with required variable mcid.

midfield_term
MIDFIELD term data table or equivalent with required variables mcid, term, and level.

...
Not used, forces later arguments to be used by name.

details
Optional flag to add columns reporting information on which the evaluation is based, default FALSE.

span
Optional numeric scalar, number of years to define timely completion. Values that are 100 span" (sched_span) are commonly used. Default 6 years.

sched_span
Optional numeric scalar, the number of years an institution officially schedules for completing a program. Default 4 years.

Value

A data.table with the following properties:

Rows are not modified.

Column timely_term is added with an option to add columns term_i, level_i, and adj_span.

Grouping structures are not preserved.

Details

The basic heuristic starts with span number of years for each student (default 6 years). The span for students admitted at a higher level than first year are reduced by one year for each full year the student is assumed to have completed.

For example, a student admitted at the second-year level is assumed to have completed one year of a program, so their span is reduced by one year. Similarly, spans are reduced by two years for students admitted at the 3rd-year level and by three years for students admitted at the fourth-year level.

The adjusted span of years is added to their starting term; the result is the timely completion term reported in the timely_term column.

The timely completion term is used in two evaluations: filtering for data sufficiency (see add_data_sufficiency()) and assessing completion for timeliness (add_completion_timely()).

If details is TRUE, additional column(s) that support the finding are returned as well. Here the extra columns are the student's initial (admission) term term_i, initial level level_i, and the adjusted span adj_span.

Existing columns with the same names as the added columns are overwritten.

See also

Examples

# Using the toy data sets
DT <- toy_student[1:10, .(mcid)]
add_timely_term(DT, midfield_term = toy_term)

#>            mcid timely_term
#>  1: MID25783939       20053
#>  2: MID25784402       20103
#>  3: MID25805538       20091
#>  4: MID25808099       19923
#>  5: MID25816437       20113
#>  6: MID25826223       19913
#>  7: MID25828870       19933
#>  8: MID25831839       20123
#>  9: MID25839453       20111
#> 10: MID25840802       19951

# Add details on which the timely term is based
add_timely_term(DT, midfield_term = toy_term, details = TRUE)

#>            mcid term_i      level_i adj_span timely_term
#>  1: MID25783939  20031    04 Senior        3       20053
#>  2: MID25784402  20081    04 Senior        3       20103
#>  3: MID25805538  20043 02 Sophomore        5       20091
#>  4: MID25808099  19871  01 Freshman        6       19923
#>  5: MID25816437  20071 02 Sophomore        5       20113
#>  6: MID25826223  19885    04 Senior        3       19913
#>  7: MID25828870  19881  01 Freshman        6       19933
#>  8: MID25831839  20081 02 Sophomore        5       20123
#>  9: MID25839453  20083    04 Senior        3       20111
#> 10: MID25840802  19913    03 Junior        4       19951

# Define timely completion as 200% of scheduled span (8 years)
add_timely_term(DT, midfield_term = toy_term, span = 8)

#>            mcid timely_term
#>  1: MID25783939       20073
#>  2: MID25784402       20123
#>  3: MID25805538       20111
#>  4: MID25808099       19943
#>  5: MID25816437       20133
#>  6: MID25826223       19933
#>  7: MID25828870       19953
#>  8: MID25831839       20143
#>  9: MID25839453       20131
#> 10: MID25840802       19971

# Optional arguments (after ...) must be named
add_timely_term(DT, toy_term, details = TRUE, span = 6)

#>            mcid term_i      level_i adj_span timely_term
#>  1: MID25783939  20031    04 Senior        3       20053
#>  2: MID25784402  20081    04 Senior        3       20103
#>  3: MID25805538  20043 02 Sophomore        5       20091
#>  4: MID25808099  19871  01 Freshman        6       19923
#>  5: MID25816437  20071 02 Sophomore        5       20113
#>  6: MID25826223  19885    04 Senior        3       19913
#>  7: MID25828870  19881  01 Freshman        6       19933
#>  8: MID25831839  20081 02 Sophomore        5       20123
#>  9: MID25839453  20083    04 Senior        3       20111
#> 10: MID25840802  19913    03 Junior        4       19951