Stickiness case data

An example of program stickiness data grouped by student race and sex.

case_stickiness

Format

A data frame with columns:

program

Instructional programs selected for study.

race

Student race and sex from midfieldstudents data.

sex

Student sex from midfieldstudents data.

ever

Number of students ever enrolled in a program. Grouped and summarized from midfieldterms data.

grad

Number of students graduating from the program. Grouped and summarized from midfielddegrees data.

stick

Program stickiness: the ratio of the number of students graduating from a program to the number ever enrolled in the program.

Examples

case_stickiness

#> # A tibble: 32 x 6
#>   program           race     sex     ever  grad stick
#>   <chr>             <chr>    <chr>  <dbl> <dbl> <dbl>
#> 1 Civil Engineering Asian    Female    22    11 0.5  
#> 2 Civil Engineering Asian    Male      54    24 0.444
#> 3 Civil Engineering Black    Female    71    21 0.296
#> 4 Civil Engineering Black    Male     132    39 0.295
#> 5 Civil Engineering Hispanic Female    16     5 0.312
#> 6 Civil Engineering Hispanic Male      85    30 0.353
#> 7 Civil Engineering White    Female   445   192 0.431
#> 8 Civil Engineering White    Male    1494   708 0.474
#> # ... with 24 more rows