Classification of Instructional Programs (CIP) data

A dataset of program codes and names of academic fields of study from the Classification of Instructional Programs (CIP) standardized by the US Department of Education.

The CIP taxonomy is organized on three levels: a 2-digit series, a 4-digit series, and a 6-digit series. This dataset has 48 unique 2-digit codes, 411 unique 4-digit codes, and 1543 unique 6-digit codes.

The MIDFIELD taxonomy includes three entries not in the standard CIP data, "NonIPEDS - Undecided/Unspecified" encoded as "99", "9999", and "999999".

cip

Format

A data frame (tibble) with 2002 observations and 2 variables. Each observation is a unique program code and name.

code

A program code of 2, 4, or 6 digits. A character variable.

name

The name of an instructional program. A character variable.

Source

https://engineering.purdue.edu/MIDFIELD