Classification of instructional programs (CIP) data

A dataset of codes and names for 1584 instructional programs organized on three levels: a 2-digit series, a 4-digit series, and a 6-digit series.

cip

Format

A tidy data frame (tibble) with 1546 observations and 6 variables. All variables are characters. An observation is a unique program.

cip2

An instructional program's 2-digit code, representing "the most general groupings of related programs" (US National Center for Education Statistics).

cip2name

Name of a program at the 2-digit level.

cip4

An instructional program's 4-digit code, representing "intermediate groupings of programs that have comparable content and objectives".

cip4name

Name of a program at the 4-digit level.

cip6

An instructional program's 6-digit code, representing "specific instructional programs".

cip6name

Name of a program at the 6-digit level.

Source

Based on the 2010 codes curated by the US National Center for Education Statistics (https://nces.ed.gov/ipeds/cipcode).

Details

6-digit codes are used in the midfieldr datasets (midfieldstudents, midfieldcourses, midfieldterms, and midfielddegrees) to encode program information as reported by the member institutions. In addition to the standard codes and names, the midfieldr taxonomy includes three non-IPEDS codes:

14XXXX

First-Year Engineering (FYE).

14YYYY

Engineering-Focused Curricula (pseudo-FYE). For instances in which a student is eligible for admission to the institution but has not been admitted to an engineering program. Some institutions without FYE programs use such a designation in lieu of undecided or unspecified.

999999

Undecided or Unspecified. For instances in which institutions reported no program information or that students were not enrolled in a program.

Examples

# View the CIP data
cip

#> # A tibble: 1,584 x 6
#>    cip2  cip2name               cip4  cip4name        cip6  cip6name            
#>    <chr> <chr>                  <chr> <chr>           <chr> <chr>               
#>  1 01    Agriculture, Agricult~ 0100  Agriculture, G~ 0100~ Agriculture, General
#>  2 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agricultural Busine~
#>  3 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agribusiness, Agric~
#>  4 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agricultural Econom~
#>  5 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Farm, Farm and Ranc~
#>  6 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agricultural, Farm ~
#>  7 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agricultural Busine~
#>  8 01    Agriculture, Agricult~ 0101  Agricultural B~ 0101~ Agricultural Busine~
#>  9 01    Agriculture, Agricult~ 0102  Agricultural M~ 0102~ Agricultural Mechan~
#> 10 01    Agriculture, Agricult~ 0102  Agricultural M~ 0102~ Agricultural Power ~
#> # ... with 1,574 more rows

# View the programs at the top level of the taxonomy
unique(cip[, 1:2])

#> # A tibble: 46 x 2
#>    cip2  cip2name                                                 
#>    <chr> <chr>                                                    
#>  1 01    Agriculture, Agricultural Operations and Related Sciences
#>  2 03    Natural Resources and Conservation                       
#>  3 04    Architecture and Related Services                        
#>  4 05    Area, Ethnic, Cultural and Gender and Group Studies      
#>  5 09    Communications, Journalism and Related Programs          
#>  6 10    Communications Technologies and Support Services         
#>  7 11    Computer and information Sciences and Support Services   
#>  8 12    Personal and Culinary Services                           
#>  9 13    Education                                                
#> 10 14    Engineering                                              
#> # ... with 36 more rows

# As the reference data argument for cip_filter()
cip_filter(series = "^1410", reference = cip)