Classification of instructional programs (CIP) data

A data set based on the US National Center for Education Statistics (NCES), Integrated Postsecondary Education Data System (IPEDS), 2020 CIP, https://nces.ed.gov/ipeds/cipcode. The data are codes and names for 1584 instructional programs organized on three levels: a 2-digit series, a 4-digit series, and a 6-digit series.

cip

Format

data.table with 1584 rows and 6 columns keyed by the 6-digit CIP code. The variables are:

cip6

character 6-digit code representing "specific instructional programs" (US National Center for Education Statistics)

cip6name

character program name at the 6-digit level

cip4

character 4-digit code (the first 4 digits of cip6) representing "intermediate groupings of programs that have comparable content and objectives"

cip4name

character program name at the 4-digit level

cip2

character 2-digit code (the first 2 digits of cip6) representing "the most general groupings of related programs"

cip2name

character program name at the 2-digit level

Details

The midfielddata taxonomy includes three non-IPEDS codes:

14XXXX

First-Year Engineering (FYE). First-Year Engineering students are neither undecided nor undeclared. Their institutions admitted them as engineering students. We just don't know what their engineering starting major would have been had they not been required to enroll in the FYE program.

14YYYY

Engineering-Focused Curricula (pseudo-FYE). For instances in which a student is eligible for admission to the institution but has not been admitted to an engineering program. Some institutions without FYE programs use such a designation in lieu of undecided or unspecified.

999999

Undecided or Unspecified. For instances in which institutions reported no program information or that students were not enrolled in a program.

See also

Examples

# overview
str(cip)

#> Classes 'data.table' and 'data.frame':	1584 obs. of  6 variables:
#>  $ cip2    : chr  "01" "01" "01" "01" ...
#>  $ cip2name: chr  "Agriculture, Agricultural Operations and Related Sciences" "Agriculture, Agricultural Operations and Related Sciences" "Agriculture, Agricultural Operations and Related Sciences" "Agriculture, Agricultural Operations and Related Sciences" ...
#>  $ cip4    : chr  "0100" "0101" "0101" "0101" ...
#>  $ cip4name: chr  "Agriculture, General" "Agricultural Business and Management" "Agricultural Business and Management" "Agricultural Business and Management" ...
#>  $ cip6    : chr  "010000" "010101" "010102" "010103" ...
#>  $ cip6name: chr  "Agriculture, General" "Agricultural Business and Management, General" "Agribusiness, Agricultural Business Operations" "Agricultural Economics" ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

# filter_by_text() argument
filter_by_text(cip, keep_text = "^1410")

#>    cip2    cip2name cip4                                               cip4name
#> 1:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 2:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 3:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#> 4:   14 Engineering 1410 Electrical, Electronics and Communications Engineering
#>      cip6                                                      cip6name
#> 1: 141001        Electrical, Electronics and Communications Engineering
#> 2: 141003                                 Laser and Optical Engineering
#> 3: 141004                                Telecommunications Engineering
#> 4: 141099 Electrical, Electronics and Communications Engineering, Other