Filter CIP data

Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.

cip_filter(series = NULL, data = NULL)

Arguments

series
The conditions used to filter the data. A character vector (or coercible to one) of any combination of 2, 4, or 6-digit CIP codes. If NULL, data is returned unchanged.

data
A data frame of character variables (or coercible to one). Required variables are CIP2, CIP4, and CIP6. If NULL, data is the midfieldr cip dataset.

Value

A data frame: data filtered by the conditions in series.

Details

The purpose of cip_filter() is to obtain the CIP codes of specific academic programs one wishes to study. The codes returned are used to identify students in the midfieldr datasets (student, course, term, degree)

Examples

cip_filter(series = c("490101", "490205"))

#>   CIP2                                     CIP2Name CIP4
#> 1   49 Transportation and Materials Moving Services 4901
#> 2   49 Transportation and Materials Moving Services 4902
#>                         CIP4Name   CIP6
#> 1    Air Transportation Services 490101
#> 2 Ground Transportation Services 490205
#>                                                          CIP6Name
#> 1 Aeronautics/Aviation/Aerospace Science and Technology - General
#> 2               Truck and Bus Driver/Commercial Vehicle Operation

cip_filter(series = c("^4901", "^4902"))

#>    CIP2                                     CIP2Name CIP4
#> 1    49 Transportation and Materials Moving Services 4901
#> 2    49 Transportation and Materials Moving Services 4901
#> 3    49 Transportation and Materials Moving Services 4901
#> 4    49 Transportation and Materials Moving Services 4901
#> 5    49 Transportation and Materials Moving Services 4901
#> 6    49 Transportation and Materials Moving Services 4901
#> 7    49 Transportation and Materials Moving Services 4901
#> 8    49 Transportation and Materials Moving Services 4902
#> 9    49 Transportation and Materials Moving Services 4902
#> 10   49 Transportation and Materials Moving Services 4902
#> 11   49 Transportation and Materials Moving Services 4902
#> 12   49 Transportation and Materials Moving Services 4902
#> 13   49 Transportation and Materials Moving Services 4902
#>                          CIP4Name   CIP6
#> 1     Air Transportation Services 490101
#> 2     Air Transportation Services 490102
#> 3     Air Transportation Services 490104
#> 4     Air Transportation Services 490105
#> 5     Air Transportation Services 490106
#> 6     Air Transportation Services 490109
#> 7     Air Transportation Services 490199
#> 8  Ground Transportation Services 490202
#> 9  Ground Transportation Services 490205
#> 10 Ground Transportation Services 490206
#> 11 Ground Transportation Services 490207
#> 12 Ground Transportation Services 490208
#> 13 Ground Transportation Services 490299
#>                                                           CIP6Name
#> 1  Aeronautics/Aviation/Aerospace Science and Technology - General
#> 2            Airline/Commercial/Professional Pilot and Flight Crew
#> 3                        Aviation/Airway Management and Operations
#> 4                                           Air Traffic Controller
#> 5                                         Airline Flight Attendant
#> 6                                                Flight Instructor
#> 7                              Air Transportation Services - Other
#> 8     Construction/Heavy Equipment/Earthmoving Equipment Operation
#> 9                Truck and Bus Driver/Commercial Vehicle Operation
#> 10                                  Mobil Crane Operation/Operator
#> 11                                    Flagging and Traffic Control
#> 12                             Railroad and Railway Transportation
#> 13                         Vehicle and Equipment Operators - Other

cip_filter(series = c("^49", "^99"))

#>    CIP2                                     CIP2Name CIP4
#> 1    49 Transportation and Materials Moving Services 4901
#> 2    49 Transportation and Materials Moving Services 4901
#> 3    49 Transportation and Materials Moving Services 4901
#> 4    49 Transportation and Materials Moving Services 4901
#> 5    49 Transportation and Materials Moving Services 4901
#> 6    49 Transportation and Materials Moving Services 4901
#> 7    49 Transportation and Materials Moving Services 4901
#> 8    49 Transportation and Materials Moving Services 4902
#> 9    49 Transportation and Materials Moving Services 4902
#> 10   49 Transportation and Materials Moving Services 4902
#> 11   49 Transportation and Materials Moving Services 4902
#> 12   49 Transportation and Materials Moving Services 4902
#> 13   49 Transportation and Materials Moving Services 4902
#> 14   49 Transportation and Materials Moving Services 4903
#> 15   49 Transportation and Materials Moving Services 4903
#> 16   49 Transportation and Materials Moving Services 4903
#> 17   49 Transportation and Materials Moving Services 4903
#> 18   49 Transportation and Materials Moving Services 4999
#> 19   99             NonIPEDS - Undecided/Unspecified 9999
#>                                              CIP4Name   CIP6
#> 1                         Air Transportation Services 490101
#> 2                         Air Transportation Services 490102
#> 3                         Air Transportation Services 490104
#> 4                         Air Transportation Services 490105
#> 5                         Air Transportation Services 490106
#> 6                         Air Transportation Services 490109
#> 7                         Air Transportation Services 490199
#> 8                      Ground Transportation Services 490202
#> 9                      Ground Transportation Services 490205
#> 10                     Ground Transportation Services 490206
#> 11                     Ground Transportation Services 490207
#> 12                     Ground Transportation Services 490208
#> 13                     Ground Transportation Services 490299
#> 14                     Marine Transportation Services 490303
#> 15                     Marine Transportation Services 490304
#> 16                     Marine Transportation Services 490309
#> 17                     Marine Transportation Services 490399
#> 18 Transportation and Materials Moving Services/Other 499999
#> 19                   NonIPEDS - Undecided/Unspecified 999999
#>                                                           CIP6Name
#> 1  Aeronautics/Aviation/Aerospace Science and Technology - General
#> 2            Airline/Commercial/Professional Pilot and Flight Crew
#> 3                        Aviation/Airway Management and Operations
#> 4                                           Air Traffic Controller
#> 5                                         Airline Flight Attendant
#> 6                                                Flight Instructor
#> 7                              Air Transportation Services - Other
#> 8     Construction/Heavy Equipment/Earthmoving Equipment Operation
#> 9                Truck and Bus Driver/Commercial Vehicle Operation
#> 10                                  Mobil Crane Operation/Operator
#> 11                                    Flagging and Traffic Control
#> 12                             Railroad and Railway Transportation
#> 13                         Vehicle and Equipment Operators - Other
#> 14                                              Commercial Fishing
#> 15                               Diver/Professional and Instructor
#> 16                          Marine Science/Merchant Marine Officer
#> 17                          Marine Transportation Services - Other
#> 18            Transportation and Materials Moving Services - Other
#> 19                                NonIPEDS - Undecided/Unspecified

cip_filter(series = c("^54", "^4901", "490205"))

#>    CIP2                                     CIP2Name CIP4
#> 1    49 Transportation and Materials Moving Services 4901
#> 2    49 Transportation and Materials Moving Services 4901
#> 3    49 Transportation and Materials Moving Services 4901
#> 4    49 Transportation and Materials Moving Services 4901
#> 5    49 Transportation and Materials Moving Services 4901
#> 6    49 Transportation and Materials Moving Services 4901
#> 7    49 Transportation and Materials Moving Services 4901
#> 8    49 Transportation and Materials Moving Services 4902
#> 9    54                                      History 5401
#> 10   54                                      History 5401
#> 11   54                                      History 5401
#> 12   54                                      History 5401
#> 13   54                                      History 5401
#> 14   54                                      History 5401
#> 15   54                                      History 5401
#> 16   54                                      History 5401
#> 17   54                                      History 5401
#>                          CIP4Name   CIP6
#> 1     Air Transportation Services 490101
#> 2     Air Transportation Services 490102
#> 3     Air Transportation Services 490104
#> 4     Air Transportation Services 490105
#> 5     Air Transportation Services 490106
#> 6     Air Transportation Services 490109
#> 7     Air Transportation Services 490199
#> 8  Ground Transportation Services 490205
#> 9                         History 540101
#> 10                        History 540102
#> 11                        History 540103
#> 12                        History 540104
#> 13                        History 540105
#> 14                        History 540106
#> 15                        History 540107
#> 16                        History 540108
#> 17                        History 540199
#>                                                           CIP6Name
#> 1  Aeronautics/Aviation/Aerospace Science and Technology - General
#> 2            Airline/Commercial/Professional Pilot and Flight Crew
#> 3                        Aviation/Airway Management and Operations
#> 4                                           Air Traffic Controller
#> 5                                         Airline Flight Attendant
#> 6                                                Flight Instructor
#> 7                              Air Transportation Services - Other
#> 8                Truck and Bus Driver/Commercial Vehicle Operation
#> 9                                                History - General
#> 10                                American History (United States)
#> 11                                                European History
#> 12                History and Philosophy of Science and Technology
#> 13              Public/Applied History and Archival Administration
#> 14                                                   Asian History
#> 15                                                Canadian History
#> 16                                                Military History
#> 17                                                 History - Other

cip_filter(series = seq(540102, 540108))

#>   CIP2 CIP2Name CIP4 CIP4Name   CIP6
#> 1   54  History 5401  History 540102
#> 2   54  History 5401  History 540103
#> 3   54  History 5401  History 540104
#> 4   54  History 5401  History 540105
#> 5   54  History 5401  History 540106
#> 6   54  History 5401  History 540107
#> 7   54  History 5401  History 540108
#>                                             CIP6Name
#> 1                   American History (United States)
#> 2                                   European History
#> 3   History and Philosophy of Science and Technology
#> 4 Public/Applied History and Archival Administration
#> 5                                      Asian History
#> 6                                   Canadian History
#> 7                                   Military History