Prepare FYE data for multiple imputation

Constructs a data frame of Student Unit Records (SURs) of First-Year Engineering (FYE) programs and conditions the data for later use as an input to the mice R package for multiple imputation. Sets up three variables as predictors (institution, race, and sex) and one variable to be predicted (program CIP code) with NA values to be imputed. Requires MIDFIELD student and term data frames in the environment.

Usage

condition_fye(midfield_student, midfield_term, ..., fye_codes = NULL)

Arguments

midfield_student

Data frame of Student Unit Record (SUR) student observations, keyed by student ID. Required variables are mcid, race, and sex.

midfield_term

Data frame of SUR term observations keyed by student ID. Required variables are mcid, institution, term, and cip6.

...

Not used, forces later arguments to be used by name.

fye_codes

Optional character vector of 6-digit CIP codes to identify FYE programs, default 140102. Codes must be 6-digit strings of numbers; regular expressions are prohibited. Non-engineering codes---those that do not start with 14---are ignored.

Value

A data.table conditioned for later use as an input to the mice R package for multiple imputation. The data have the following properties:

One row for every FYE student.

Grouping structures are not preserved.

Columns returned:

mcid

Character, anonymized student identifier. Returned as-is.

race

Factor, race/ethnicity as self-reported by the student. An imputation predictor variable.

sex

Factor, sex as self-reported by the student. An imputation predictor variable.

institution

Factor, anonymized institution name. An imputation predictor variable.

cip6

Factor, 6-digit CIP code of a student's known, post-FYE engineering program or NA representing missing values to be imputed.

Details

At some US institutions, First-Year Engineering (FYE) programs comprise a common first year curriculum that is a prerequisite for declaring a engineering major. In longitudinal studies, FYE programs can be problematic when combining their observations with those of non-FYE institutions. For example, suppose we are calculating a graduation rate for Mechanical Engineering (ME) programs. By most definitions, the students starting in ME constitute the pool; the fraction of the pool that graduates in ME is the rate. FYE starters, even if they graduate in ME, are excluded from the count because they did not start in ME.

Therefore, to include FYE students in any persistence metric requiring a degree-granting "starting" program, we have to predict the engineering program (such as Electrical Engineering, Mechanical Engineering, etc.) the FYE student would have declared had they not been required to enroll in FYE. The purpose of condition_fye() is to prepare the data for making that prediction.

Method

The method requires MIDFIELD-style student and term data tables that contain at a minimum the institutions being studied that also have FYE programs. The function extracts all terms for all FYE students and sorts them into two categories:

Students who complete an FYE and declare an engineering major. This is the easy case--at the student's first opportunity, they enrolled in an engineering program of their choosing. We use that program as the predicted starting program.

Students who, after FYE, do not declare an engineering major. This is the more complicated case---the data provide no information regarding what engineering program the student would have declared originally had the institution not required them to enroll in FYE. For these students, we treat the starting program as missing data to be predicted using multiple imputation.

In cases where students enter FYE, change programs, and re-enter FYE, only the first group of FYE terms is considered. Any programs before FYE are ignored. The first (if any) post-FYE program is identified. If the program is in engineering, the CIP is retained as their predicted starting major. If not, the CIP is replaced with NA to be treated as missing data for the imputation.

Lastly, the predictor variables (institution, race, sex) and the predicted variable (6-digit CIP code) are converted to unordered factors. The resulting data frame is ready for use as input for the mice package.

See also

Examples

# # Using toy data
# DT <- toy_student[, .(mcid, race, sex)]
# condition_fye(dframe = DT, midfield_term = toy_term)
# 
# 
# # Overwrites institution if present in dframe
# DT <- toy_student[, .(mcid, institution, race, sex)]
# condition_fye(dframe = DT, midfield_term = toy_term)
# 
# 
# # Other columns, if any, are dropped
# colnames(toy_student)
# colnames(condition_fye(toy_student, toy_term))
# 
# 
# # Optional argument permits multiple CIP codes for FYE
# condition_fye(DT, toy_term, fye_codes = c("140101", "140102"))