Condition multiway data for graphing

Transform a data frame such that two categorical variables are factors with levels ordered for display in a Cleveland "multiway dot plot," where the ordering of the panels and rows are crucial to the perception of effects. In multiway data---as defined by Cleveland (1993)---there is a single quantitative value (or response) for every combination of levels of two categorical variables. Typically the quantitative column is used to order the levels of the two categorical columns.

Usage

condition_multiway(dframe, x, yy, ..., order_by = NULL, param_col = NULL)

Arguments

dframe

Data frame with at least one numeric variable and two categorical variables (character or factor).

x

Character scalar, the quoted name of the numeric response variable.

yy

Character vector, the quoted names of the two categorical variables.

...

Not used, forces later arguments to be used by name.

order_by

Optional character scalar (in quotes) assigning the method for ordering the levels of the categorical variables. The following values are possible:

"median" (default) Orders by the median of values in the quantitative column grouped by category. This is the order recommended by Cleveland when the quantitative variable is an integer count or frequency.

"mean" Orders by the mean of values in the quantitative column grouped by category.

"sum" Orders by the sum of values in the quantitative column grouped by category. Useful also when the quantitative variable is a count or frequency.

"percent" Orders by ratios computed by category. Used when the quantitative response variable is the ratio (in percent) of two columns of integer counts (frequencies). The counts are summed by category to obtain grouped percentages used to order the levels of the categorical variables. Requires two parameter columns be identified in param_col and assumes that the parameter with the larger column sum is the denominator of the ratio.

"alphabet" Orders the levels of the categorical variables alphabetically. Rarely useful for perceiving effects, but can be useful for value look up.

param_col

Optional character vector with the names (in quotes) of the columns used as parameters in the order_by method. Currently supports "percent" method only---expects two column names (in any order) of the integer count columns used to construct the percentage response column.

Value

A data.table with the following properties:

Rows are not modified.

The quantitative column is not modified.

The two categorical columns are factors with levels ordered by the method selected.

Grouping structures are not preserved.

Details

In the multiway dot plot, there are panels, the individual dot plots of the display, and there are levels, the rows of each panel. One category is encoded by the panels; the other by the rows. All panels have the same quantitative scale on the x-axis and the same organization of category levels on the y-axis. Panels and rows are ordered so that ordering scheme increases in "graph order", that is, increases from left to right and from bottom to top.

Note that "multiway" in our context refers to the data structure and graph design defined by Cleveland, not to the methods of analysis described by Kroonenberg (2008).

References

Cleveland WS (1993). Visualizing Data. Hobart Press, Summit, NJ.

Kroonenberg PM (2008). Applied Multiway Data Analysis. Wiley, Hoboken, NJ.

See also