Filter CIP data

Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.

filter_cip(series = NULL, data = NULL)

Arguments

series
The conditions used to filter the data. A character vector (or coercible to one) of 2, 4, or 6-digit CIP codes.

data
A data frame of CIP codes and program names at the 2, 4, and 6-digit levels: CIP2, PRGM2, CIP4, PRGM4, CIP6, and PRGM6. The default is the midfieldr cip dataset. All variables are characters.

Value

A data frame: data filtered by the conditions in series.

Details

In the midfieldr datasets (student, course, term, degree) a 6-digit CIP code identifies the program in which a student is enrolled at matriculation, in a term, and at graduation. The purpose of filter_cip() is to obtain the 6-digit codes of the programs one wishes to study.

Examples

filter_cip(series = c("490101", "490205"))

#> # A tibble: 2 x 6
#>    CIP2                                        PGRM2  CIP4
#>   <chr>                                        <chr> <chr>
#> 1    49 Transportation and Materials Moving Services  4901
#> 2    49 Transportation and Materials Moving Services  4902
#> # ... with 3 more variables: PGRM4 <chr>, CIP6 <chr>, PGRM6 <chr>

filter_cip(series = c("4901", "4902"))

#> # A tibble: 13 x 6
#>    CIP2                                        PGRM2  CIP4
#>   <chr>                                        <chr> <chr>
#> 1    49 Transportation and Materials Moving Services  4901
#> 2    49 Transportation and Materials Moving Services  4901
#> 3    49 Transportation and Materials Moving Services  4901
#> 4    49 Transportation and Materials Moving Services  4901
#> 5    49 Transportation and Materials Moving Services  4901
#> 6    49 Transportation and Materials Moving Services  4901
#> # ... with 7 more rows, and 3 more variables: PGRM4 <chr>, CIP6 <chr>,
#> #   PGRM6 <chr>

filter_cip(series = c("49", "99"))

#> # A tibble: 19 x 6
#>    CIP2                                        PGRM2  CIP4
#>   <chr>                                        <chr> <chr>
#> 1    49 Transportation and Materials Moving Services  4901
#> 2    49 Transportation and Materials Moving Services  4901
#> 3    49 Transportation and Materials Moving Services  4901
#> 4    49 Transportation and Materials Moving Services  4901
#> 5    49 Transportation and Materials Moving Services  4901
#> 6    49 Transportation and Materials Moving Services  4901
#> # ... with 13 more rows, and 3 more variables: PGRM4 <chr>, CIP6 <chr>,
#> #   PGRM6 <chr>

filter_cip(series = c("54", "4901", "490205"))

#> # A tibble: 17 x 6
#>    CIP2                                        PGRM2  CIP4
#>   <chr>                                        <chr> <chr>
#> 1    49 Transportation and Materials Moving Services  4901
#> 2    49 Transportation and Materials Moving Services  4901
#> 3    49 Transportation and Materials Moving Services  4901
#> 4    49 Transportation and Materials Moving Services  4901
#> 5    49 Transportation and Materials Moving Services  4901
#> 6    49 Transportation and Materials Moving Services  4901
#> # ... with 11 more rows, and 3 more variables: PGRM4 <chr>, CIP6 <chr>,
#> #   PGRM6 <chr>

filter_cip(series = seq(540102, 540108))

#> # A tibble: 7 x 6
#>    CIP2   PGRM2  CIP4   PGRM4   CIP6
#>   <chr>   <chr> <chr>   <chr>  <chr>
#> 1    54 History  5401 History 540102
#> 2    54 History  5401 History 540103
#> 3    54 History  5401 History 540104
#> 4    54 History  5401 History 540105
#> 5    54 History  5401 History 540106
#> 6    54 History  5401 History 540107
#> # ... with 1 more rows, and 1 more variables: PGRM6 <chr>