Subset rows by matching values in shared key columns

Subset rows of one data frame by values of a key column that match values of a similar key column in a second data frame. Columns are not subset unless selected in an optional argument.

Usage

filter_match(dframe, match_to, by_col, ..., select = NULL)

Arguments

dframe

Data frame to be subset and returned. Must contain the key column named in by_col.

match_to

Data frame with key column values to be matched to. The only column used is the required key column named in by_col.

by_col

Character scalar, name of the key column. Values in the key column must be character strings.

...

Not used, force later arguments to be used by name.

select

Optional character vector of dframe column names to retain, default all columns.

Value

A data.table with the following properties:

Rows with matching values.

All columns or those specified by select.

Grouping structures are not preserved.

Details

Two data frames are input. The match_to data frame is subset to retain its key column only. The result is merged with dframe in an inner-join, returning the rows of dframe with values matching the key values in match_to.

Column subsetting occurs after the inner join, so the key column does not have to be included in the vector of column names in select.

See also

Examples

# Start with a toy sample of student (built-in data set)
head(toy_student)
#>           mcid   institution              transfer hours_transfer  race    sex
#> 1: MID25783939 Institution M First-Time in College             NA White Female
#> 2: MID25784402 Institution M First-Time in College             NA White   Male
#> 3: MID25805538 Institution M First-Time in College             NA White Female
#> 4: MID25808099 Institution M First-Time in College             NA White Female
#> 5: MID25816437 Institution M First-Time in College             NA White   Male
#> 6: MID25826223 Institution M   First-Time Transfer             NA White Female
nrow(toy_student)
#> [1] 100

# Filter by matching IDs of graduates
student_example <- filter_match(toy_student,
  match_to = toy_degree,
  by_col = "mcid"
)
head(student_example)
#>           mcid   institution              transfer hours_transfer  race    sex
#> 1: MID25783939 Institution M First-Time in College             NA White Female
#> 2: MID25808099 Institution M First-Time in College             NA White Female
#> 3: MID25826223 Institution M   First-Time Transfer             NA White Female
#> 4: MID25828870 Institution M First-Time in College             NA White   Male
#> 5: MID25840802 Institution M First-Time in College             NA White Female
#> 6: MID25845841 Institution M First-Time in College             NA White Female
# Note change in number of rows
nrow(student_example)
#> [1] 64

# Same filter and select 3 columns only
student_example <- filter_match(toy_student,
  match_to = toy_degree,
  by_col = "mcid",
  select = c("mcid", "race", "sex")
)
head(student_example)
#>           mcid  race    sex
#> 1: MID25783939 White Female
#> 2: MID25808099 White Female
#> 3: MID25826223 White Female
#> 4: MID25828870 White   Male
#> 5: MID25840802 White Female
#> 6: MID25845841 White Female
nrow(student_example)
#> [1] 64

# Filter term data for engineering program codes (start with "14")
term_example <- toy_term[grepl("^14", cip6), ]
head(term_example)
#>           mcid   institution  term   cip6        level hours_term
#> 1: MID25784402 Institution M 20081 141901    04 Senior         19
#> 2: MID25816437 Institution M 20071 140601 02 Sophomore         15
#> 3: MID25828870 Institution M 19881 140102  01 Freshman         17
#> 4: MID25828870 Institution M 19883 140102 02 Sophomore         14
#> 5: MID25828870 Institution M 19891 140102 02 Sophomore         14
#> 6: MID25828870 Institution M 19893 149999    03 Junior         17
nrow(term_example)
#> [1] 78

# Filter student by matching IDs of engineering students
student_example <- filter_match(toy_student,
  match_to = term_example,
  by_col = "mcid",
  select = c("mcid", "institution", "transfer", "sex")
)
head(student_example)
#>           mcid   institution              transfer    sex
#> 1: MID25784402 Institution M First-Time in College   Male
#> 2: MID25816437 Institution M First-Time in College   Male
#> 3: MID25828870 Institution M First-Time in College   Male
#> 4: MID25839453 Institution M First-Time in College Female
#> 5: MID25845841 Institution M First-Time in College Female
#> 6: MID25846316 Institution M First-Time in College   Male
nrow(student_example)
#> [1] 26

# The 'by_col' column does not have to be included in the `select` columns
student_example <- filter_match(toy_student,
  match_to = term_example,
  by_col = "mcid",
  select = c("institution", "transfer", "sex")
)
head(student_example)
#>      institution              transfer    sex
#> 1: Institution M First-Time in College   Male
#> 2: Institution M First-Time in College   Male
#> 3: Institution M First-Time in College   Male
#> 4: Institution M First-Time in College Female
#> 5: Institution M First-Time in College Female
#> 6: Institution M First-Time in College   Male
nrow(student_example)
#> [1] 26