Subset rows by matching values in a common key column

Subset rows of one data frame by values of a key variable that match values of the same key variable in a second data frame.

filter_match(dframe, match_to, by_col, ..., select = NULL)

Arguments

dframe
Data frame to be subset and returned, column identified in by_col required.

match_to
Data frame to match to, column identified in by_col required

by_col
Character scalar, name of the key column.

...
Not used, force later arguments to be used by name.

select
Character vector of dframe column names to retain, default all columns.

Value

A data.table with the following properties:

Unique rows with matching values.

All columns or those specified by select.

Grouping structures are not preserved.

Details

Two data frames are input. The match_to data frame is subset to retain its key column only. The result is merged with dframe in an inner-join, returning the rows of dframe with values matching the key values in match_to.

All columns in dframe are returned unless a subset is given in the select argument. Column subsetting occurs after the inner join, so the key column does not have to be included in the vector of column names in select. The final operation is to subset for unique rows.

See also

Examples

# TBD