Filter degree data for graduating students

Filter academic degree data to find all students graduating from specified programs.

grad_filter(
  data = NULL,
  codes = NULL,
  ...,
  id = "id",
  cip6 = "cip6",
  term = "term_degree"
)

Arguments

data
Data frame of student IDs, academic terms, and CIP codes, default midfielddegrees.

codes
Atomic character vector of 6-digit CIP codes specifying the programs to filter by.

...
Not used for values, forces later arguments to bind by name

id
The column name in quotes of the student ID variable in data. Default is "id".

cip6
The column name in quotes of the 6-digit CIP code variable in data. Default is "cip6".

term
The column name in quotes of the term variable in data. Default is "term".

Value

Data frame with character variables for student ID and program CIP

Details

The midfielddata package must be installed to provide the midfielddegrees data frame as the default value for the data argument. Optionally, a data frame having the same structure as midfielddegrees may be used as the value for the data argument.

The codes argument is an atomic character vector of 6-digit CIP codes used to filter data and extract student IDs and terms.

The function returns a subset of data with every unique combination of student ID and CIP of the program(s) in which they earned their first degree(s). Only the student ID and program CIP are returned; other variables in data are quietly dropped.

Optional arguments. Student ID variables in the midfielddata data sets are named "id". If your data frames use a different name, you can either 1) rename your variables to "id", or 2) use the optional id argument to pass the alternate variable name to the function. The same is true for the cip6 and term variables.

See also

Examples

library("midfielddata")
(grad_filter(codes = "540104"))

#> # A tibble: 24 x 2
#>    id          cip6  
#>    <chr>       <chr> 
#>  1 MID26360410 540104
#>  2 MID26361391 540104
#>  3 MID26362301 540104
#>  4 MID26362581 540104
#>  5 MID26365694 540104
#>  6 MID26365705 540104
#>  7 MID26366791 540104
#>  8 MID26367223 540104
#>  9 MID26368839 540104
#> 10 MID26383658 540104
#> # ... with 14 more rows