Graph functions

Provides a set of customized layers for graphs created with for ggplot2.

expon_scale_x_log10()

midfield_theme()

Value

Modifies an existing graph

Details

expon_scale_x_log10()

Applies the usual scale_x_log10() function with the scale marked with powers of ten in exponential form and the logarithmic minor grid lines shown.

midfield_theme()

Applies theme_minimal() with some additional edits.