Tools for Student Records Research

A package for investigating a data sample from the MIDFIELD database. The data comprise student records from registrars at participating US universities, including demographic, term, course, and degree information for 97,640 undergraduate students from 1987 to 2016.

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Details

This package provides functions to access, manipulate, and graph student record data. To illustrate the workflow, we would use these functions in sequence to compute and display the "stickiness" metric:

cip_filter() to select programs from the CIP data

cip_label() to add custom program labels for grouping

gather_ever() to gather students ever enrolled

gather_grad() to gather graduating students

race_sex_join() to join student sex and race

seplyr::group_summarize() to count students

dplyr::join() and mutate() to compute stickiness

multiway_order() to prepare the results for graphing

ggplot2::ggplot() to graph the metric

To learn about these and other functions in detail, start with the vignettes: browseVignettes(package = "midfieldr").

Data

Instructional program data is provided in cip, a data frame of names and codes of academic programs from the 2010 IPEDS Classification of Instructional Programs (CIP) by the US Department of Education.

Student record data is provided in the companion data package, midfielddata. Data are provided in four datasets: midfieldcourses, midfielddegrees, midfieldstudents, and midfieldterms.