Undergraduate Student Records for Longitudinal Research

A package for investigating a data sample from the MIDFIELD database. The data comprise student records from registrars at participating US universities, including demographic, term, course, and degree information for 97,640 undergraduate students from 1987 to 2016.

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Details

In our examples illustrating how to use midfieldr tools and datasets, the findings are generally conditioned by instructional program, student sex, and student race/ethnicity. Sex and race/ethnicity categories are limited to those reported by the US institutions contributing to the MIDFIELD database.

Student record data is provided in four data packages: midfieldstudents, midfieldcourses, midfieldterms, and midfielddegrees.

Data

cip

A dataset of names and codes of academic programs. Each of the 1544 observations is a program described by a 6-digit Classification of Instructional Programs (CIP) code defined by the US Department of Education in 2010.

midfieldstudents

In the midfieldstudents data package. A tidy data frame with 97,640 observations and 15 demographic variables. Each observation is a unique student. Occupies 19 Mb of memory.

midfieldcourses

In the midfieldcourses data package. A tidy data frame with 3.5 M observations and 12 academic course variables. Each observation is one course in one term for one student. Occupies 348 Mb of memory.

midfieldterms

In the midfieldterms data package. A tidy data frame with 729,014 observations and 13 academic term variables. Each observation is one term for one student. Occupies 82 Mb of memory.

midfielddegrees

In the midfielddegrees data package. A tidy data frame with 97,640 observations and 5 graduation variables. Each observation is a unique student. Of these students, 47,499 earned their degrees. Occupies 10 Mb of memory.

Functions

This package provides functions to access, manipulate, and graph the student record data.

cip_filter()

Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.

join_demographics()

Add variables ETHNIC and SEX from the student dataset to a data frame.