Undergraduate Student Records For Longitudinal Research

A package for investigating a sample from the MIDFIELD database. The data comprise student records from registrars at participating US universities, including demographic, term, course, and degree information for 165,000 undergraduate students from 1990 to 2016.

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Details

In our examples illustrating how to use midfieldr tools and datasets, the findings are generally disaggregated or conditioned by instructional program and student sex and race/ethnicity. Sex and race/ethnicity categories are limited to those reported by the US institutions contributing to the MIDFIELD database.

Datasets are provided with midfieldr plus three data packages: midfieldstudent, midfieldterm, and midfieldcourse.

Data

student

midfieldstudent package. Demographic data for 165,000 students. Each observation is a unique student.

term

midfieldterm package. Academic term data for 165,000 students. Each observation is one term for one student, yielding 1.1 M observations.

course

midfieldcourse package. Academic course data for 165,000 students. Each observation is one course for one student, yielding 5.4 M observations.

degree

midfieldstudent package. Graduation data for 90,000 students receiving degrees. Each observation is a unique student.

cip

midfieldr package. A dataset of program codes and names of academic fields of study. Each of the 1552 observations is one program at the CIP 6-digit level.

Functions

midfieldr provides functions to access, manipulate, and graph the student record data.

cip_filter()

Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.