midfieldr: Tools for Student Records Research

For investigating undergraduate student records in the MIDFIELD database that comprises data from registrar's at participating US institutions. Designed to work with MIDFIELD-structured data, providing tools for accessing and manipulating student, term, course, and degree information. Data are available for 97,640 undergraduate students from 1987 to 2016.

Details

A typical midfieldr workflow might include:

filter_text() search the CIP data set for program codes

prepare_fye_mi() predict starting programs of first-year-engineering (FYE) students

subset_feasible() subset students for 6-year completion feasibility

prepare_multiway() condition multiway data for graphing

Operations such subsetting, joining, and grouping and summarizing are performed using one's preferred tools in base R or more advanced systems such as dplyr or data.table. The examples use base R for data carpentry and ggplot2 for graphs.

Data

Instructional program data is provided in cip, a data frame of names and codes of academic programs from the 2010 IPEDS Classification of Instructional Programs (CIP) by the US Department of Education.

Student record data is provided in a companion data package midfielddata. Data are provided in four data sets: midfieldstudents, midfieldcourses, midfieldterms, and midfielddegrees.

For more information

vignette(package = 'midfieldr')

website midfieldr

website midfielddata

website MIDFIELD