Tools for Studying MIDFIELD Student Unit Record Data in R

The goal of midfieldr is to provide tools for working with MIDFIELD data, a resource of longitudinal, de-identified, individual student unit records.

Details

MIDFIELD contains individual Student Unit Record (SUR) data for 1.7M students at 33 US institutions (as of June 2021). MIDFIELD is large enough to permit grouping and summarizing by multiple characteristics, enabling researchers to examine student characteristics (race/ethnicity, sex, prior achievement) and curricular pathways (including coursework and major) by institution and over time.

A proportionate stratified sample of these data (for practice) is available in midfielddata, an R data package with longitudinal student unit-records for 98,000 undergraduates at 12 institutions from 1987--2016 organized in four data tables:

student

course

term

degree

The tools in midfieldr work equally well with the practice data in midfielddata and the research data in MIDFIELD.

This work is supported by a grant from the US National Science Foundation (EEC 1545667).