midfieldr: Tools for Student Records Research

The *Multiple-Institution Database for Investigating Engineering Longitudinal Development* (MIDFIELD) is a partnership of US higher education institutions with engineering programs. MIDFIELD contains registrar's data for 1.7M undergraduates in all majors at 19 institutions from 1987--2019 (Ohland and Long 2016) .

Details

The software environment comprises two R packages:

midfieldr: An R package providing functions specialized for manipulating MIDFIELD data to examine the intersectionality of race/ethnicity, sex, and discipline in persistence metrics such as stickiness (retention by a discipline) and graduation rate.

midfielddata: An R package with practice data for users to learn about student record analysis using R. However, these data are not suitable for drawing inferences about student performance, i.e., not for research.

midfielddata provides a proportionate stratified random sample of the MIDFIELD research data. The sampling strata are institution, cip4 (the first four digits of the 6-digit CIP code), transfer status, race/ethnicity, and sex. Contains four data tables: student, course, term, and degree keyed by an anonymized student ID.

Complete MIDFIELD data suitable for student-records research are available to researchers under the following conditions:

Your institutional IRB has granted approval for your project to study students using MIDFIELD. At most institutions, the use of MIDFIELD data for research is in the IRB "Exempt" category, but institutional practices vary.

Each researcher using the data signs a letter of confidentiality describing the guidelines for how the data may be reported.

The research data and practice data have the same structure with the same variable names, though some research variables are omitted from the practice data. Thus R scripts written for the practice data should generally work with the research data.

References

Ohland MW, Long RA (2016). “The Multiple-Institution Database for Investigating Engineering Longitudinal Development: An experiential case study of data sharing and reuse.” Advances in Engineering Education, 5(2), 398--404. http://advances.asee.org/wp-content/uploads/vol05/issue02/Papers/AEE-18-Ohland.pdf.