Prepare multiway data for graphing

Transforms a data frame such that two categorical variables become factors with their levels ordered by medians of the quantitative variable.

multiway_order(x, medians = FALSE)

Arguments

x
A data frame with one numerical variable and two categorical variables.

medians
A logical value indicating whether or not the medians computed to order the factors should be included in the data frame returned. The default is FALSE.

Value

A data frame with its two categorical variables converted to factors and ordered by increasing medians of the numerical variable. The medians of the quantitative variable for each category are added to the data frame if argument medians is TRUE.

Details

In multiway data, there is a single quantitative value (or response) for every combination of levels of two categorical variables. multiway_order() converts the categorical variables to factors and orders the factors by increasing median values of the levels of the categorical variables.

Examples

# construct a data frame with one numeric and two character variables
set.seed(20150531)
cat1 <- rep(c("Art", "History", "Math"), each = 2)
cat2 <- rep(c("Male", "Female"), times = 3)
x    <- round(runif(6, min = 0.1, max = 0.7), 3)
df1  <- tibble::data_frame(cat1, cat2, x)
df1

#> # A tibble: 6 x 3
#>   cat1    cat2       x
#>   <chr>   <chr>  <dbl>
#> 1 Art     Male   0.355
#> 2 Art     Female 0.497
#> 3 History Male   0.449
#> 4 History Female 0.398
#> 5 Math    Male   0.674
#> 6 Math    Female 0.257

# convert characters to factors ordered by medians
df2 <- multiway_order(df1)
df2 <- dplyr::arrange(df2, cat1, cat2)
df2

#> # A tibble: 6 x 3
#>   cat1    cat2       x
#>   <fct>   <fct>  <dbl>
#> 1 History Female 0.398
#> 2 History Male   0.449
#> 3 Art     Female 0.497
#> 4 Art     Male   0.355
#> 5 Math    Female 0.257
#> 6 Math    Male   0.674

# include the medians in the output
df3 <- multiway_order(df1, medians = TRUE)
df3 <- dplyr::arrange(df3, cat1, cat2)
df3

#> # A tibble: 6 x 5
#>   cat1    cat2       x med_cat1 med_cat2
#>   <fct>   <fct>  <dbl>    <dbl>    <dbl>
#> 1 History Female 0.398    0.424    0.398
#> 2 History Male   0.449    0.424    0.449
#> 3 Art     Female 0.497    0.426    0.398
#> 4 Art     Male   0.355    0.426    0.449
#> 5 Math    Female 0.257    0.466    0.398
#> 6 Math    Male   0.674    0.466    0.449