Order levels of multiway data

Transform a data frame such that two categorical variables are factors with levels ordered by medians of the quantitative variable. Rows and columns of the data frame are otherwise unaffected.

order_multiway(data = NULL)

Arguments

data
data frame of multiway data

Value

data.frame with the following properties:

Rows are not modified

Quantitative column is not modified

Categorical columns are factors with levels ordered by median quantitative values

Grouping structures, if any, are not preserved

Data frame extensions tbl or data.table are preserved

Details

In multiway data, there is a single quantitative value (or response) for every combination of levels of two categorical variables. order_multiway() converts the columns of categorical variables to factors and orders the factor levels by increasing medians of the quantitative response variable.

The data argument, therefore, must have three columns only. One column is a quantitative variable (type numeric) and two columns are categorical variables (type character or factor).

See also

Examples

catg1 <- rep(c("urban", "rural", "suburb", "village"), each = 2)
catg2 <- rep(c("men", "women"), times = 4)
value <- c(0.22, 0.14, 0.43, 0.58, 0.81, 0.46, 0.15, 0.20)
mw_df <- data.frame(catg1, catg2, value)
order_multiway(mw_df)

#>     catg1 catg2 value
#> 1   urban   men  0.22
#> 2   urban women  0.14
#> 3   rural   men  0.43
#> 4   rural women  0.58
#> 5  suburb   men  0.81
#> 6  suburb women  0.46
#> 7 village   men  0.15
#> 8 village women  0.20