Order categorical variables of multiway data

Transform a data frame such that two independent categorical variables are factors with levels ordered for display in a multiway dot plot. Multiway data comprise a single quantitative value (or response) for every combination of levels of two categorical variables. The ordering of the rows and panels is crucial to the perception of effects (Cleveland, 1993).

Usage

order_multiway(
  dframe,
  quantity,
  categories,
  ...,
  method = NULL,
  ratio_of = NULL
)

Arguments

dframe

Data frame with one numeric variable and two categorical variables of class character or factor. Two additional numeric columns required when using the "percent" ordering method.

quantity

Character, name (in quotes) of the single multiway quantitative variable

categories

Character, vector of names (in quotes) of the two multiway categorical variables

...

Not used, forces later arguments to be used by name.

method

Character, “median” (default) or “percent”, method of ordering the levels of the categories. The median method computes the medians of the quantitative column grouped by category. The percent method computes percentages based on the same ratio underlying the quantitative percentage variable except grouped by category.

ratio_of

Character vector with the names (in quotes) of the numerator and denominator columns that produced the quantitative variable, required when method is "percent". Names can be in any order; the algorithm assumes that the parameter with the larger column sum is the denominator of the ratio.

Value

A data.table with the following properties:

Rows are not modified.

Grouping structures are not preserved.

The columns specified by categories are converted to factors and ordered. Other columns are not modified.

Two columns are added. Caution! An existing column with the same name as one of the added columns is silently overwritten.

The names of the added columns incorporate the names of the multiway variables. Columns added:

CATEGORY_median columns (when ordering method is "median")

Numeric. Two columns of medians of the quantitative variable grouped by the categorical variables. The CATEGORY placeholder in the column name is replaced by a category name from the categories argument. For example, suppose categories = c("program", "people") and method = "median". The two new column names would be program_median and people_median.

CATEGORY_QUANTITY columns (when ordering method is "percent")

Numeric. Two columns of percentages based on the same ratio that produces the quantitative variable except grouped by the categorical variables. The CATEGORY placeholder in the column name is replaced by a category name from the categories argument; the QUANTITY placeholder is replaced by the quantitative variable name in the quantity argument. For example, suppose categories = c("program", "people"), and quantity = "grad_rate", and method = "percent". The two new column names would be program_grad_rate and people_grad_rate.

Details

In our context, "multiway" refers to the data structure and graph design defined by Cleveland (1993), not to the methods of analysis described by Kroonenberg (2008).

Multiway data comprise three variables: a categorical variable of m levels; a second independent categorical variable of n levels; and a quantitative variable (or response) of length mn that cross-classifies the categories, that is, there is a value of the response for each combination of levels of the two categorical variables.

In a multiway dot plot, one category is encoded by the panels, the second category is encoded by the rows of each panel, and the quantitative variable is encoded along identical horizontal scales.

References

Cleveland WS (1993). Visualizing Data. Hobart Press, Summit, NJ.

Kroonenberg PM (2008). Applied Multiway Data Analysis. Wiley, Hoboken, NJ.

Examples

# Subset of built-in data set
dframe <- study_results[program == "EE" | program == "ME"]
dframe[, people := paste(race, sex)]
#>     program    sex            race ever grad stick                 people
#>  1:      EE Female           Asian   36   14  38.9           Asian Female
#>  2:      EE Female           Black  145   58  40.0           Black Female
#>  3:      EE Female   International    8    3  37.5   International Female
#>  4:      EE Female          Latine   14    6  42.9          Latine Female
#>  5:      EE Female Native American    3    0   0.0 Native American Female
#>  6:      EE Female   Other/Unknown    8    3  37.5   Other/Unknown Female
#>  7:      EE Female           White  173   55  31.8           White Female
#>  8:      EE   Male           Asian  189   86  45.5             Asian Male
#>  9:      EE   Male           Black  287   97  33.8             Black Male
#> 10:      EE   Male   International   70   35  50.0     International Male
#> 11:      EE   Male          Latine   63   22  34.9            Latine Male
#> 12:      EE   Male Native American    8    1  12.5   Native American Male
#> 13:      EE   Male   Other/Unknown   27    9  33.3     Other/Unknown Male
#> 14:      EE   Male           White 1227  509  41.5             White Male
#> 15:      ME Female           Asian   22   13  59.1           Asian Female
#> 16:      ME Female           Black   75   23  30.7           Black Female
#> 17:      ME Female   International    3    1  33.3   International Female
#> 18:      ME Female          Latine   10    4  40.0          Latine Female
#> 19:      ME Female Native American    5    1  20.0 Native American Female
#> 20:      ME Female   Other/Unknown    8    4  50.0   Other/Unknown Female
#> 21:      ME Female           White  261  109  41.8           White Female
#> 22:      ME   Male           Asian  118   58  49.2             Asian Male
#> 23:      ME   Male           Black  202   65  32.2             Black Male
#> 24:      ME   Male   International   36   16  44.4     International Male
#> 25:      ME   Male          Latine   76   29  38.2            Latine Male
#> 26:      ME   Male Native American   14    4  28.6   Native American Male
#> 27:      ME   Male   Other/Unknown   43   20  46.5     Other/Unknown Male
#> 28:      ME   Male           White 1776  918  51.7             White Male
#>     program    sex            race ever grad stick                 people
dframe[, c("race", "sex") := NULL]
#>     program ever grad stick                 people
#>  1:      EE   36   14  38.9           Asian Female
#>  2:      EE  145   58  40.0           Black Female
#>  3:      EE    8    3  37.5   International Female
#>  4:      EE   14    6  42.9          Latine Female
#>  5:      EE    3    0   0.0 Native American Female
#>  6:      EE    8    3  37.5   Other/Unknown Female
#>  7:      EE  173   55  31.8           White Female
#>  8:      EE  189   86  45.5             Asian Male
#>  9:      EE  287   97  33.8             Black Male
#> 10:      EE   70   35  50.0     International Male
#> 11:      EE   63   22  34.9            Latine Male
#> 12:      EE    8    1  12.5   Native American Male
#> 13:      EE   27    9  33.3     Other/Unknown Male
#> 14:      EE 1227  509  41.5             White Male
#> 15:      ME   22   13  59.1           Asian Female
#> 16:      ME   75   23  30.7           Black Female
#> 17:      ME    3    1  33.3   International Female
#> 18:      ME   10    4  40.0          Latine Female
#> 19:      ME    5    1  20.0 Native American Female
#> 20:      ME    8    4  50.0   Other/Unknown Female
#> 21:      ME  261  109  41.8           White Female
#> 22:      ME  118   58  49.2             Asian Male
#> 23:      ME  202   65  32.2             Black Male
#> 24:      ME   36   16  44.4     International Male
#> 25:      ME   76   29  38.2            Latine Male
#> 26:      ME   14    4  28.6   Native American Male
#> 27:      ME   43   20  46.5     Other/Unknown Male
#> 28:      ME 1776  918  51.7             White Male
#>     program ever grad stick                 people
data.table::setcolorder(dframe, c("program", "people"))

# Class before ordering
class(dframe$program)
#> [1] "character"
class(dframe$people)
#> [1] "character"

# Class and levels after ordering
mw1 <- order_multiway(dframe, 
                      quantity = "stick", 
                      categories = c("program", "people"))
class(mw1$program)
#> [1] "factor"
levels(mw1$program)
#> [1] "EE" "ME"
class(mw1$people)
#> [1] "factor"
levels(mw1$people)
#>  [1] "Native American Female" "Native American Male"   "Black Male"            
#>  [4] "Black Female"           "International Female"   "Latine Male"           
#>  [7] "White Female"           "Other/Unknown Male"     "Latine Female"         
#> [10] "Other/Unknown Female"   "White Male"             "International Male"    
#> [13] "Asian Male"             "Asian Female"          

# Display category medians 
mw1
#>     program                 people stick ever grad program_median people_median
#>  1:      EE           Asian Female  38.9   36   14           37.5         49.00
#>  2:      EE           Black Female  40.0  145   58           37.5         35.35
#>  3:      EE   International Female  37.5    8    3           37.5         35.40
#>  4:      EE          Latine Female  42.9   14    6           37.5         41.45
#>  5:      EE Native American Female   0.0    3    0           37.5         10.00
#>  6:      EE   Other/Unknown Female  37.5    8    3           37.5         43.75
#>  7:      EE           White Female  31.8  173   55           37.5         36.80
#>  8:      EE             Asian Male  45.5  189   86           37.5         47.35
#>  9:      EE             Black Male  33.8  287   97           37.5         33.00
#> 10:      EE     International Male  50.0   70   35           37.5         47.20
#> 11:      EE            Latine Male  34.9   63   22           37.5         36.55
#> 12:      EE   Native American Male  12.5    8    1           37.5         20.55
#> 13:      EE     Other/Unknown Male  33.3   27    9           37.5         39.90
#> 14:      EE             White Male  41.5 1227  509           37.5         46.60
#> 15:      ME           Asian Female  59.1   22   13           40.9         49.00
#> 16:      ME           Black Female  30.7   75   23           40.9         35.35
#> 17:      ME   International Female  33.3    3    1           40.9         35.40
#> 18:      ME          Latine Female  40.0   10    4           40.9         41.45
#> 19:      ME Native American Female  20.0    5    1           40.9         10.00
#> 20:      ME   Other/Unknown Female  50.0    8    4           40.9         43.75
#> 21:      ME           White Female  41.8  261  109           40.9         36.80
#> 22:      ME             Asian Male  49.2  118   58           40.9         47.35
#> 23:      ME             Black Male  32.2  202   65           40.9         33.00
#> 24:      ME     International Male  44.4   36   16           40.9         47.20
#> 25:      ME            Latine Male  38.2   76   29           40.9         36.55
#> 26:      ME   Native American Male  28.6   14    4           40.9         20.55
#> 27:      ME     Other/Unknown Male  46.5   43   20           40.9         39.90
#> 28:      ME             White Male  51.7 1776  918           40.9         46.60
#>     program                 people stick ever grad program_median people_median

# Existing factors (if any) are re-ordered
mw2 <- dframe
mw2$program <- factor(mw2$program, levels = c("ME", "EE"))

# Levels before conditioning
levels(mw2$program) 
#> [1] "ME" "EE"

# Levels after conditioning
mw2 <- order_multiway(dframe, 
                      quantity = "stick", 
                      categories = c("program", "people"))
levels(mw2$program) 
#> [1] "EE" "ME"

# Ordering using percent method
order_multiway(dframe, 
               quantity = "stick", 
               categories = c("program", "people"), 
               method = "percent", 
               ratio_of = c("grad", "ever"))
#>     program                 people grad ever stick program_stick people_stick
#>  1:      EE           Asian Female   14   36  38.9          39.8         46.6
#>  2:      EE           Black Female   58  145  40.0          39.8         36.8
#>  3:      EE   International Female    3    8  37.5          39.8         36.4
#>  4:      EE          Latine Female    6   14  42.9          39.8         41.7
#>  5:      EE Native American Female    0    3   0.0          39.8         12.5
#>  6:      EE   Other/Unknown Female    3    8  37.5          39.8         43.8
#>  7:      EE           White Female   55  173  31.8          39.8         37.8
#>  8:      EE             Asian Male   86  189  45.5          39.8         46.9
#>  9:      EE             Black Male   97  287  33.8          39.8         33.1
#> 10:      EE     International Male   35   70  50.0          39.8         48.1
#> 11:      EE            Latine Male   22   63  34.9          39.8         36.7
#> 12:      EE   Native American Male    1    8  12.5          39.8         22.7
#> 13:      EE     Other/Unknown Male    9   27  33.3          39.8         41.4
#> 14:      EE             White Male  509 1227  41.5          39.8         47.5
#> 15:      ME           Asian Female   13   22  59.1          47.8         46.6
#> 16:      ME           Black Female   23   75  30.7          47.8         36.8
#> 17:      ME   International Female    1    3  33.3          47.8         36.4
#> 18:      ME          Latine Female    4   10  40.0          47.8         41.7
#> 19:      ME Native American Female    1    5  20.0          47.8         12.5
#> 20:      ME   Other/Unknown Female    4    8  50.0          47.8         43.8
#> 21:      ME           White Female  109  261  41.8          47.8         37.8
#> 22:      ME             Asian Male   58  118  49.2          47.8         46.9
#> 23:      ME             Black Male   65  202  32.2          47.8         33.1
#> 24:      ME     International Male   16   36  44.4          47.8         48.1
#> 25:      ME            Latine Male   29   76  38.2          47.8         36.7
#> 26:      ME   Native American Male    4   14  28.6          47.8         22.7
#> 27:      ME     Other/Unknown Male   20   43  46.5          47.8         41.4
#> 28:      ME             White Male  918 1776  51.7          47.8         47.5
#>     program                 people grad ever stick program_stick people_stick