Case-study stickiness results

Longitudinal stickiness results for Civil, Electrical, Industrial, and Mechanical Engineering programs from midfielddata. Results are keyed by program, race/ethnicity, and sex. Used in the case study developed in the vignettes.

study_stickiness

Format

data.table with 38 rows and 6 columns keyed by program, race/ethnicity, and sex.

program

Character, program name assigned for grouping and summarizing and for display in graphs and tables.

race

Character, race/ethnicity as self-reported by the student, e.g., Asian, Black, Hispanic/LatinX, etc.

sex

Character, sex as self-reported by the student, values are Female, Male, and Unknown.

ever

Numerical, number of students ever enrolled in the program, after subsetting for data sufficiency and small populations.

grad

Numerical, number of students completing a program after subsetting for data sufficiency and accounting for timely completion.

stick

Numerical, program stickiness, the ratio of the number of students completing a program to the number of students ever enrolled in the program.

Details

Longitudinal stickiness is the ratio of the number of students graduating from a program to the number of students ever enrolled in the program over the time span of available data.

Data are subset to satisfy the data sufficiency criterion and some graduates are reclassified as non-completers to satisfy the timely completion criterion. Data are also subset to remove groups with 10 or fewer members in the ever column to reduce the risk of identification.

See also