Degree data for examples

A small subset by row and column of the MIDFIELD degree table for use in package examples.

toy_degree

Format

data.table with 64 rows and 5 columns keyed by student ID.

mcid

Character, anonymized student ID.

institution

Character, anonymized institution name, e.g., Institution A, Institution B, etc.

term

Character, academic year and term in which a student completes their program, format YYYYT.

cip6

Character, 6-digit CIP code of program in which a student earns a degree.

degree

Character, type of degree awarded, e.g., Bachelor's Degree, Bachelor of Arts, Bachelor of Science, etc.

See also